% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.anylist}
\alias{aggregate.anylist}
\title{Aggregate \link[spatstat.geom]{anylist}}
\usage{
\method{aggregate}{anylist}(x, by, ...)
}
\arguments{
\item{x}{an \link[spatstat.geom]{anylist}}

\item{by}{\link[base]{factor}, of same \link[base]{length} as \code{x}}
}
\description{
Aggregate \link[spatstat.geom]{anylist}
}
\keyword{internal}
