% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.groupedHyperframe.R
\name{aggregate.groupedHyperframe}
\alias{aggregate.groupedHyperframe}
\title{Aggregate \link{groupedHyperframe}}
\usage{
\method{aggregate}{groupedHyperframe}(x, by, fun = pmean, ...)
}
\arguments{
\item{x}{a \link{groupedHyperframe}}

\item{by}{a one-sided \link[stats]{formula}}

\item{fun}{\link[base]{function}, aggregation method,
currently supports
\code{\link[=pmean]{pmean()}}, \code{\link[=pmedian]{pmedian()}}, \link[base]{pmax}, and \link[base]{pmin}.}

\item{...}{additional parameters}
}
\value{
Function \code{\link[=aggregate.groupedHyperframe]{aggregate.groupedHyperframe()}} returns a \link[spatstat.geom]{hyperframe}.
}
\description{
Aggregate \link{groupedHyperframe}
}
\details{
Function \code{\link[=aggregate.groupedHyperframe]{aggregate.groupedHyperframe()}} checks \code{by} against \code{attr(x,'group')}.
}
\keyword{internal}
