% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.hyperframe.R
\name{aggregate.hyperframe}
\alias{aggregate.hyperframe}
\title{Aggregate Hyper Data Frame}
\usage{
\method{aggregate}{hyperframe}(x, by, ...)
}
\arguments{
\item{x}{\link[spatstat.geom]{hyperframe}}

\item{by}{two-sided \link[stats]{formula},
whose right-hand-side contains only the regular-column names of the input \code{x}}

\item{...}{additional parameters of the function \link[stats]{aggregate.data.frame},
\emph{except for} \code{simplify}}
}
\value{
The \code{S3} method \code{\link[=aggregate.hyperframe]{aggregate.hyperframe()}} returns a \link[spatstat.geom]{hyperframe}.
}
\description{
To \link[stats]{aggregate} a \link[spatstat.geom]{hyperframe}.
}
