% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.vectorlist}
\alias{aggregate.vectorlist}
\title{Aggregate \code{vectorlist}}
\usage{
\method{aggregate}{vectorlist}(x, by, fun = pmean, ...)
}
\arguments{
\item{x}{a \code{vectorlist}}

\item{by}{\link[base]{factor}, of same \link[base]{length} as \code{x}}

\item{fun}{\link[base]{function}, aggregation method,
currently supports functions
\code{\link[=pmean]{pmean()}}, \code{\link[=pmedian]{pmedian()}}, \link[base]{pmax}, and \link[base]{pmin}.}

\item{...}{additional parameters, currently of no use}
}
\description{
Aggregate \code{vectorlist}
}
\keyword{internal}
