% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate2.R
\name{aggregate2}
\alias{aggregate2}
\title{Aggregate, An Alternative Way}
\usage{
aggregate2(x, by, ...)
}
\arguments{
\item{x}{a \link[base]{data.frame}}

\item{by}{a two-sided \link[stats]{formula}}

\item{...}{additional parameters of the function \link[stats]{aggregate.data.frame}}
}
\description{
An alternative aggregation function,
to avoid the \link[base]{cbind}-operation in the function \link[stats]{aggregate.formula}.
}
\note{
The function \link[stats]{aggregate.data.frame} is the workhorse of
the function \link[stats]{aggregate.formula}.

The function \link[spatstat.geom]{as.hyperframe.data.frame}
is \strong{designed} to handle the \link[base]{list}-columns
returned by the function \link[stats]{aggregate}.
}
\keyword{internal}
