% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate2hyper.R
\name{aggregate2hyper}
\alias{aggregate2hyper}
\title{Aggregate to (Grouped) Hyper Data Frame}
\usage{
aggregate2hyper(x, by, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{by}{\link[stats]{formula}}

\item{...}{additional parameters, currently of no use}
}
\description{
Aggregate to (Grouped) Hyper Data Frame
}
\keyword{internal}
