% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_FUN.R
\name{aggregate_FUN}
\alias{aggregate_FUN}
\alias{unique_or_identity}
\title{Handy \code{FUN}ctions for \link[stats]{aggregate}}
\usage{
unique_or_identity(x)
}
\arguments{
\item{x}{a column in a \link[base]{data.frame}, i.e., a \link[base]{vector}, a \link[survival]{Surv} object, etc.}
}
\value{
The function \code{\link[=unique_or_identity]{unique_or_identity()}} returns an R object.
}
\description{
Handy functions to be passed to the parameter \code{FUN} of the function \link[stats]{aggregate}
and/or \code{\link[=aggregate2]{aggregate2()}}.
}
