% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_marks.R
\name{aggregate_marks}
\alias{aggregate_marks}
\alias{aggregate_marks.ppp}
\alias{aggregate_marks.ppplist}
\alias{aggregate_marks.hyperframe}
\title{\link[stats]{aggregate} the \link[spatstat.geom]{marks} of \link[spatstat.geom]{ppp}-Related Object}
\usage{
aggregate_marks(x, by, FUN, expr, ..., vectorize = FALSE)

\method{aggregate_marks}{ppp}(x, by, FUN, expr, ..., vectorize = FALSE)

\method{aggregate_marks}{ppplist}(x, ...)

\method{aggregate_marks}{hyperframe}(x, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{by, FUN, ...}{additional parameters of the function \link[stats]{aggregate.data.frame}}

\item{expr}{\link[base]{expression}, only used when \code{markformat.ppp(x)} is \code{'dataframe'}
\strong{and} the parameter \code{by} is missing}

\item{vectorize}{\link[base]{logical} scalar (default value \code{FALSE}), whether to convert the return from
The function \link[stats]{aggregate.data.frame} into a \link[base]{vector}}
}
\description{
To \link[stats]{aggregate} the \link[spatstat.geom]{marks} of a \link[spatstat.geom]{ppp}-related object, e.g.,
\itemize{
\item{\link[spatstat.geom]{ppp.object}}
\item{a \code{ppplist}}
\item{the \link[spatstat.geom]{ppp}-hypercolumn of a \link[spatstat.geom]{hyperframe}}
}
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
