% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_marks.R
\name{append_marks_set}
\alias{append_marks_set}
\alias{append_marks<-}
\alias{append_marks<-.ppp}
\alias{append_marks<-.tess}
\alias{append_marks<-.psp}
\title{Append to Existing \link[spatstat.geom]{marks}}
\usage{
append_marks(x) <- value

\method{append_marks}{ppp}(x) <- value

\method{append_marks}{tess}(x) <- value

\method{append_marks}{psp}(x) <- value
}
\arguments{
\item{x}{currently only \link[spatstat.geom]{ppp.object} is supported}

\item{value}{a \link[base]{factor} or \link[base]{numeric} \link[base]{vector}}
}
\value{
The \code{S3} method dispatch \code{\link{append_marks<-.ppp}} returns a \link[spatstat.geom]{ppp.object}.
}
\description{
Append an additional \code{mark} to existing \link[spatstat.geom]{marks}.
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
