% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvlist.R
\name{as.fvlist}
\alias{as.fvlist}
\title{Convert a \link[stats]{listof} \link[spatstat.explore]{fv.object}s into \code{'fvlist'}}
\usage{
as.fvlist(X, data.name)
}
\arguments{
\item{X}{a \link[stats]{listof} \link[spatstat.explore]{fv.object}s}

\item{data.name}{\link[base]{character} scalar, name of \code{X}, for console message output}
}
\value{
The function \code{\link[=as.fvlist]{as.fvlist()}} returns an \link[base]{invisible} \link[base]{list}.
}
\description{
Convert a \link[stats]{listof} \link[spatstat.explore]{fv.object}s into \code{'fvlist'}
}
\keyword{internal}
