% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvlist.R
\name{as.list.fvlist}
\alias{as.list.fvlist}
\title{\code{\link[=as.list.fvlist]{as.list.fvlist()}}}
\usage{
\method{as.list}{fvlist}(x, ...)
}
\arguments{
\item{x}{\code{fvlist}}

\item{...}{additional parameters, currently of no use}
}
\description{
Converts an \code{fvlist} to a simple \link[base]{list}
}
\keyword{internal}
