% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_spatial.R
\name{as.matrix.pairwise_cor_spatial}
\alias{as.matrix.pairwise_cor_spatial}
\title{Convert \link{pairwise_cor_spatial} to \link[base]{matrix}}
\usage{
\method{as.matrix}{pairwise_cor_spatial}(x, ...)
}
\arguments{
\item{x}{a \link{pairwise_cor_spatial}}

\item{...}{additional parameters, currently not in use}
}
\value{
The function \code{\link[=as.matrix.pairwise_cor_spatial]{as.matrix.pairwise_cor_spatial()}} returns a \link[base]{matrix}.
}
\description{
Convert \link{pairwise_cor_spatial} to \link[base]{matrix}
}
\keyword{internal}
