% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_.R
\name{attr_}
\alias{attr_}
\alias{attr_.anylist}
\alias{attr_.hyperframe}
\title{A Batch Process of \link[base]{attr}}
\usage{
attr_(x, which, exact = TRUE)

\method{attr_}{anylist}(x, which, exact = TRUE)

\method{attr_}{hyperframe}(x, which, exact = TRUE)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{which, exact}{parameters of the function \link[base]{attr}}
}
\description{
A Batch Process of \link[base]{attr}
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
