% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aug4gam.R
\name{aug4gam}
\alias{aug4gam}
\alias{aug4gam.data.frame}
\alias{aug4gam.hyperframe}
\title{Augment Hypercolumn(s) for \link[mgcv]{gam}}
\usage{
aug4gam(data, ...)

\method{aug4gam}{data.frame}(data, ...)

\method{aug4gam}{hyperframe}(data, ...)
}
\arguments{
\item{data}{\link[spatstat.geom]{hyperframe}}

\item{...}{additional parameters, currently of no use}
}
\description{
Augment Hypercolumn(s) for \link[mgcv]{gam}
}
\details{
All \link[base]{numeric} \link[base]{vector} hypercolumns will be augmented.
}
\author{
Tingting Zhan, Erjia Cui
}
