% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibentry.R
\name{bib}
\alias{bib}
\alias{bib_btaf430}
\alias{bib_groupedHyperframe}
\alias{bib_hypergam}
\title{Citations of \pkg{groupedHyperframe} and its Applications}
\usage{
bib_btaf430()

bib_groupedHyperframe()

bib_hypergam()
}
\description{
Citations of \pkg{groupedHyperframe} and its Applications
}
\keyword{internal}
