% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.groupedHyperframe.R
\name{cbind.groupedHyperframe}
\alias{cbind.groupedHyperframe}
\title{\code{cbind.groupedHyperframe()}}
\usage{
\method{cbind}{groupedHyperframe}(...)
}
\arguments{
\item{...}{see the function \link[spatstat.geom]{cbind.hyperframe},
the first element must be \code{'groupedHyperframe'}}
}
\description{
\code{cbind.groupedHyperframe()}
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
