% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Defunct.R
\name{cli_defunct_}
\alias{cli_defunct_}
\alias{cli_RPubs_}
\alias{cli_doi_}
\title{Defunct Messages using Package \CRANpkg{cli}}
\usage{
cli_RPubs_(x)

cli_doi_(x)
}
\arguments{
\item{x}{\link[base]{character} scalar}
}
\description{
Defunct Messages using Package \CRANpkg{cli}
}
\keyword{internal}
