% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtrapz.R
\name{cumvtrapz}
\alias{cumvtrapz}
\alias{cumvtrapz.numeric}
\alias{cumvtrapz.fv}
\title{Cumulative Average Vertical Height of Trapezoidal Integration}
\usage{
cumvtrapz(x, ...)

\method{cumvtrapz}{numeric}(x, y, ..., rm1 = TRUE, drop = FALSE)

\method{cumvtrapz}{fv}(x, key = fvnames(x, a = ".y"), .x = fvnames(x, a = ".x"), ...)
}
\arguments{
\item{x}{\link[base]{numeric} \link[base]{vector}}

\item{...}{additional parameters of the function \link[pracma]{trapz} and \link[pracma]{cumtrapz}}

\item{y}{see function \link[pracma]{cumtrapz}}

\item{rm1}{\link[base]{logical} scalar, whether to remove the first \code{NaN}-value from
the \code{S3} generic function \code{\link[=cumvtrapz]{cumvtrapz()}} return, default \code{TRUE}}

\item{drop}{\link[base]{logical} scalar (default value is \code{FALSE}),
whether to return a \link[base]{numeric} \link[base]{vector}
instead of an \link[base]{ncol}-\code{1L} \link[base]{matrix}}

\item{key, .x}{\link[base]{character} scalars}
}
\value{
The \code{S3} generic function \code{\link[=cumvtrapz]{cumvtrapz()}} return a \link[base]{numeric} scalar.
}
\description{
Cumulative trapezoidal integration divided by \eqn{x}-domain.
}
\note{
This is a tentative thought: the prefix \code{v} stands for 'vertical'.
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
