% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette.R
\name{.ns_generic}
\alias{.ns_generic}
\title{Generic Function with Namespace}
\usage{
.ns_generic(x, backtick = TRUE, ver = FALSE)
}
\arguments{
\item{x}{\link[base]{character} scalar}

\item{backtick}{\link[base]{logical} scalar}

\item{ver}{\link[base]{logical} scalar}
}
\description{
Generic Function with Namespace
}
\examples{
.ns_generic('names<-')
.ns_generic('Math', ver = TRUE)
.ns_generic('update')

}
\keyword{internal}
