% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{.slice}
\alias{.slice}
\title{.slice}
\usage{
.slice(x, j)
}
\arguments{
\item{x}{a \link[stats]{listof} \link[base]{vector}s of same \link[base]{length}}

\item{j}{\link[base]{integer} or \link[base]{character} scalar or \link[base]{vector}}
}
\value{
Function \code{\link[=.slice]{.slice()}} returns a \link[base]{vector}
}
\description{
.slice
}
\keyword{internal}
