% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.R
\name{drop_lowest_nested}
\alias{drop_lowest_nested}
\title{Drop Lowest Nested Level}
\usage{
drop_lowest_nested(g)
}
\arguments{
\item{g}{an R \link[base]{language} object}
}
\description{
Drop Lowest Nested Level
}
\examples{
drop_lowest_nested(~ g1/g2a:g2b/g3a:g3b:g3c)
drop_lowest_nested(~ g1)
drop_lowest_nested(~ g1:g2)

}
\keyword{internal}
