% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emark_.R
\name{elig_marks}
\alias{elig_marks}
\alias{Emark_}
\alias{Emark_.ppp}
\alias{Vmark_}
\alias{Vmark_.ppp}
\alias{Kmark_}
\alias{Kmark_.ppp}
\alias{markcorr_}
\alias{markcorr_.ppp}
\alias{markvario_}
\alias{markvario_.ppp}
\alias{Gcross_}
\alias{Gcross_.ppp}
\alias{Jcross_}
\alias{Jcross_.ppp}
\alias{Kcross_}
\alias{Kcross_.ppp}
\alias{Lcross_}
\alias{Lcross_.ppp}
\alias{markconnect_}
\alias{markconnect_.ppp}
\alias{nncross_}
\alias{nncross_.ppp}
\title{On Eligible \link[spatstat.geom]{marks}}
\usage{
Emark_(X, ...)

\method{Emark_}{ppp}(X, ...)

Vmark_(X, ...)

\method{Vmark_}{ppp}(X, ...)

Kmark_(X, ...)

\method{Kmark_}{ppp}(X, ...)

markcorr_(X, ...)

\method{markcorr_}{ppp}(X, ...)

markvario_(X, ...)

\method{markvario_}{ppp}(X, ...)

Gcross_(X, ...)

\method{Gcross_}{ppp}(X, ...)

Jcross_(X, ...)

\method{Jcross_}{ppp}(X, ...)

Kcross_(X, ...)

\method{Kcross_}{ppp}(X, ...)

Lcross_(X, ...)

\method{Lcross_}{ppp}(X, ...)

markconnect_(X, ...)

\method{markconnect_}{ppp}(X, ...)

nncross_(X, ...)

\method{nncross_}{ppp}(X, ...)
}
\arguments{
\item{X}{see \strong{Usage}}

\item{...}{additional parameters of user operation}
}
\description{
On Eligible \link[spatstat.geom]{marks}
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
