% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randLabel_test.R
\name{foobar}
\alias{foobar}
\title{Random Re-Labelling Test}
\usage{
foobar(
  Y,
  fun = Kmark,
  f = function(m1, m2) {
     m1 * m2
 },
  ...
)
}
\arguments{
\item{Y}{a \link[spatstat.geom]{ppp.object}, see function \link[spatstat.explore]{envelope.ppp}}

\item{fun, f, ...}{arguments of function \link[spatstat.explore]{envelope.ppp}}
}
\description{
Random Re-Labelling Test
}
\examples{
set.seed(52); spatstat.data::anemones |>
  foobar() |>
  attr(which = 'p', exact = TRUE)

}
\keyword{internal}
