% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastLegal.R
\name{fv2theo}
\alias{fv2theo}
\alias{.illegal2theo}
\alias{.disrecommend2theo}
\alias{.illegal2theo.fv}
\alias{.disrecommend2theo.fv}
\title{Tentative Fix for Illegal Function Value in \code{fv.object}}
\usage{
.illegal2theo(X, ...)

.disrecommend2theo(X, ...)

\method{.illegal2theo}{fv}(X, key = fvnames(X, a = ".y"), .x = fvnames(X, a = ".x"), ...)

\method{.disrecommend2theo}{fv}(
  X,
  key = fvnames(X, a = ".y"),
  .x = fvnames(X, a = ".x"),
  ...
)
}
\arguments{
\item{X}{an \link[spatstat.explore]{fv.object}}

\item{key, .x}{\link[base]{character} scalars}
}
\description{
Tentative Fix for Illegal Function Value in \code{fv.object}
}
\keyword{internal}
