% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.R
\name{get_nested}
\alias{get_nested}
\title{Get Nested Groups}
\usage{
get_nested(by)
}
\arguments{
\item{by}{an R \link[base]{language} object, (nested) grouping structure}
}
\value{
The function \code{\link[=get_nested]{get_nested()}} returns a \link[base]{list} of \link[base]{language} object(s).
}
\description{
Get Nested Groups
}
\examples{
get_nested(~ g1)
get_nested(~ g1:g2)
get_nested(~ g1/g2a:g2b/g3a:g3b:g3c)

library(nlme)
nlme::Wafer |>
 getGroupsFormula() |>
 get_nested()

}
\keyword{internal}
