% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.R
\name{get_nested_factor}
\alias{get_nested_factor}
\title{Get Nested Factors}
\usage{
get_nested_factor(data, by)
}
\arguments{
\item{data}{see \emph{Usage}}

\item{by}{an R \link[base]{language} object, see function \link{get_nested}}
}
\value{
The \code{S3} generic function \code{\link[=get_nested_factor]{get_nested_factor()}} returns
a \link[base]{list} of \link[base]{factor}s.
}
\description{
Get Nested Factors
}
\examples{
nlme::Wafer |>
 get_nested_factor()

spatstat.data::osteo |>
 get_nested_factor(by = ~ id/brick)

}
