% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedHyperframe.R
\name{getGroupsFormula.groupedHyperframe}
\alias{getGroupsFormula.groupedHyperframe}
\title{Extract Grouping Formula from \link{groupedHyperframe}}
\usage{
\method{getGroupsFormula}{groupedHyperframe}(object, asList, sep)
}
\arguments{
\item{object}{a \link{groupedHyperframe}}

\item{asList, sep}{place holders for the \code{S3} generic function \link[nlme]{getGroupsFormula}.
Ignored in this \code{S3} method}
}
\value{
The \code{S3} method \code{getGroupsFormula.groupedHyperframe()} returns a one-sided \link[stats]{formula}.
}
\description{
..
}
\keyword{internal}
