% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlabelRes.R
\name{global_envelope_test_}
\alias{global_envelope_test_}
\alias{global_envelope_test_.anylist}
\alias{global_envelope_test_.hyperframe}
\title{Global Envelope Test in Bath Process}
\usage{
global_envelope_test_(x, ...)

\method{global_envelope_test_}{anylist}(x, ...)

\method{global_envelope_test_}{hyperframe}(x, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{parameters (except the first) of the function \link[GET]{global_envelope_test}}
}
\description{
Global Envelope Test in Bath Process
}
\keyword{internal}
