% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_ppp.R
\name{grouped_ppp}
\alias{grouped_ppp}
\title{Hyper Data Frame with One-and-Only-One \link[spatstat.geom]{ppp}-Hyper Column}
\usage{
grouped_ppp(
  formula,
  by,
  data,
  coords = ~x + y,
  window = owin(xrange = range(.x), yrange = range(.y)),
  ...
)
}
\arguments{
\item{formula}{\link[stats]{formula} in the format of
\code{m1+m2 ~ .},
where
\describe{
\item{\eqn{m_i}}{one or more \link[spatstat.geom]{marks}}
\item{\code{.}}{the endpoint and predictor(s) for downstream analysis}
}}

\item{by}{\link[stats]{formula}, see function \code{\link[=aggregate2hyper]{aggregate2hyper()}}}

\item{data}{a \link[base]{data.frame}}

\item{coords}{\link[stats]{formula}, variable names
of the \eqn{x}- and \eqn{y}-coordinates in \code{data}.
Default value is \code{~x+y}.}

\item{window}{an observation window \link[spatstat.geom]{owin},
default value is the \eqn{x}- and \eqn{y}-span of \code{coords} in \code{data}.}

\item{...}{additional parameters, currently not in use}
}
\value{
The function \code{\link[=grouped_ppp]{grouped_ppp()}} returns a (grouped) hyper data frame
with \emph{one-and-only-one}
\link[spatstat.geom]{ppp}-hyper column.
}
\description{
To create a (grouped) hyper data frame with one-and-only-one \link[spatstat.geom]{ppp}-hyper column.
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/nonS3/grouped_ppp_appx.html}
}
\keyword{internal}
