% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{groupedHyperframe-package}
\alias{groupedHyperframe}
\alias{groupedHyperframe-package}
\title{groupedHyperframe: Grouped Hyper Data Frame}
\description{
To aggregate a hyper data frame, defined in the package 'spatstat.geom', according to a grouping structure. To facilitate downstream analysis based on a "grouped" hyper data frame. The author has retired from academic research. Accordingly, this package should not be considered a validated tool for use in peer-reviewed publications or as the basis for grant applications. Backward compatibility with user-code published in \doi{10.1093/bioinformatics/btaf430} is not maintained in versions >= 0.4.0 (June 2026) of this package. The authors of those publications are the appropriate contacts for reproducibility inquiries.
}
\note{
This R package author has retired from academic research.

Accordingly, this package should not be considered a validated tool for use in peer-reviewed publications or as the basis for grant applications.

Backward compatibility with user-code published in the following publications is not maintained in versions \verb{>= 0.4.0} of this package. The authors of those publications are the appropriate contacts for reproducibility inquiries.

\doi{10.1093/bioinformatics/btaf430}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tingtingzhan/groupedHyperframe}
}

}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

Authors:
\itemize{
  \item Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})
}

}
