% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{groupGeneric_fvlist}
\alias{groupGeneric_fvlist}
\alias{Math.fvlist}
\alias{Summary.fvlist}
\title{\link[base]{groupGeneric} of \code{'fvlist'}}
\usage{
\method{Math}{fvlist}(x, ...)

\method{Summary}{fvlist}(..., na.rm = FALSE)
}
\arguments{
\item{x}{an \code{'fvlist'}}

\item{...}{additional parameters for the \link[base]{groupGeneric} functions}

\item{na.rm}{\link[base]{logical} scalar}
}
\value{
The function \code{\link[=Math.fvlist]{Math.fvlist()}} returns an \code{'fvlist'}.

The function \code{\link[=Summary.fvlist]{Summary.fvlist()}} ..
}
\description{
...
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
