% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{groupGeneric_ppp}
\alias{groupGeneric_ppp}
\alias{Math.ppp}
\alias{Summary.ppp}
\title{\link[base]{groupGeneric} of \link[spatstat.geom]{ppp.object}}
\usage{
\method{Math}{ppp}(x, ...)

\method{Summary}{ppp}(..., na.rm = FALSE)
}
\arguments{
\item{x}{a \link[spatstat.geom]{ppp.object}}

\item{...}{additional parameters for the \link[base]{groupGeneric} functions}

\item{na.rm}{\link[base]{logical} scalar}
}
\value{
The function \code{\link[=Math.ppp]{Math.ppp()}} returns a \link[spatstat.geom]{ppp.object}.

The function \code{\link[=Summary.ppp]{Summary.ppp()}} ..
}
\description{
...
}
\details{
The function \code{\link[=Math.ppp]{Math.ppp()}} performs \code{Math} operations on the \link[base]{numeric} \link[spatstat.geom]{marks}
of a \link[spatstat.geom]{ppp.object}.

The function \code{\link[=Summary.ppp]{Summary.ppp()}} ..
}
\keyword{internal}
