% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{groupGeneric_ppplist}
\alias{groupGeneric_ppplist}
\alias{Math.ppplist}
\alias{Summary.ppplist}
\title{\link[base]{groupGeneric} of \code{'ppplist'}}
\usage{
\method{Math}{ppplist}(x, ...)

\method{Summary}{ppplist}(..., na.rm = FALSE)
}
\arguments{
\item{x}{an \code{'ppplist'}}

\item{...}{additional parameters for the \link[base]{groupGeneric} functions}

\item{na.rm}{\link[base]{logical} scalar}
}
\value{
The function \code{\link[=Math.ppplist]{Math.ppplist()}} returns a \code{'ppplist'}.

The function \code{\link[=Summary.ppplist]{Summary.ppplist()}} ...
}
\description{
...
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
