% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastLegal.R
\name{illegal2theo}
\alias{illegal2theo}
\alias{.illegal2theo}
\alias{.illegal2theo.fv}
\alias{.illegal2theo.fvlist}
\alias{.illegal2theo.hyperframe}
\title{Tentative Fix for Illegal Function Value in \code{fv.object}}
\usage{
.illegal2theo(X, ...)

\method{.illegal2theo}{fv}(X, ...)

\method{.illegal2theo}{fvlist}(X, ...)

\method{.illegal2theo}{hyperframe}(X, ...)
}
\arguments{
\item{X}{an \link[spatstat.explore]{fv.object}}
}
\description{
Tentative Fix for Illegal Function Value in \code{fv.object}
}
\keyword{internal}
