% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.fv.R
\name{interpolation_fv}
\alias{interpolation_fv}
\alias{approxfun.fv}
\alias{splinefun.fv}
\alias{interpSpline_.fv}
\title{Interpolation of \link[spatstat.explore]{fv.object}}
\usage{
approxfun.fv(x, key = fvnames(fv, a = ".y"), .x = fvnames(fv, a = ".x"), ...)

splinefun.fv(x, key = fvnames(fv, a = ".y"), .x = fvnames(fv, a = ".x"), ...)

interpSpline_.fv(
  x,
  key = fvnames(fv, a = ".y"),
  .x = fvnames(fv, a = ".x"),
  ...
)
}
\arguments{
\item{x}{an \link[spatstat.explore]{fv.object}}

\item{key, .x}{\link[base]{character} scalars}

\item{...}{additional parameters of the functions
\link[stats]{approxfun}, \link[stats]{splinefun} and \link[splines]{interpSpline}}
}
\description{
Interpolation of \link[spatstat.explore]{fv.object}
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
