% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvlist.R
\name{is.fvlist}
\alias{is.fvlist}
\title{Inspect a \link[stats]{listof} \link[spatstat.explore]{fv.object}s}
\usage{
is.fvlist(X)
}
\arguments{
\item{X}{a \link[stats]{listof} \link[spatstat.explore]{fv.object}s}
}
\value{
The function \code{\link[=is.fvlist]{is.fvlist()}} returns a \link[base]{logical} scalar with \link[base]{attributes}
\describe{
\item{\code{attr(,'r')}}{\eqn{x}-axis, or the \eqn{r}-values}
\item{\code{attr(,'fname')}}{see explanation of this \link[base]{attributes} in function \link[spatstat.explore]{fv}}
\item{\code{attr(,'.x')}}{\code{spatstat.explore::fvnames(x, a = '.x')} returns}
\item{\code{attr(,'.y')}}{\code{spatstat.explore::fvnames(x, a = '.y')} returns}
}
}
\description{
A helper function to check the validity of a \link[stats]{listof} \link[spatstat.explore]{fv.object}s.
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
