% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_S3.R
\name{is.numeric.ppp}
\alias{is.numeric.ppp}
\title{Is \link[spatstat.geom]{marks} of \link[spatstat.geom]{ppp.object} \link[base]{numeric} ?}
\usage{
\method{is.numeric}{ppp}(x)
}
\arguments{
\item{x}{a \link[spatstat.geom]{ppp.object}}
}
\description{
Is \link[spatstat.geom]{marks} of \link[spatstat.geom]{ppp.object} \link[base]{numeric} ?
}
\keyword{internal}
