% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_vectorlist.R
\name{is.vectorlist}
\alias{is.vectorlist}
\title{Vector-List}
\usage{
is.vectorlist(
  x,
  mode = c("logical", "integer", "numeric", "double", "character")
)
}
\arguments{
\item{x}{a \link[stats]{listof}}

\item{mode}{\link[base]{character} scalar other than \code{'any'}, \code{'complex'} and '\code{raw}',
see the function \link[base]{is.vector}}
}
\value{
The function \code{\link[=is.vectorlist]{is.vectorlist()}} returns a \link[base]{logical} scalar.
}
\description{
To determine if an R object is a \link[base]{list} of
\link[base]{vector}s with the same
\link[base]{storage.mode},
\link[base]{length} and
\link[base]{attributes}.
}
\examples{
spatstat.data::Kovesi$values |>
 is.vectorlist(mode = 'character') |>
 stopifnot()
spatstat.data::Kovesi$values |>
 is.vectorlist(mode = 'numeric')
}
