% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key1val.R
\name{keyval.fv}
\alias{keyval.fv}
\title{Black-Solid-Curve in \link[spatstat.explore]{plot.fv}}
\usage{
keyval.fv(x, key = fvnames(x, a = ".y"))
}
\arguments{
\item{x}{an \link[spatstat.explore]{fv.object}}

\item{key}{\link[base]{character} scalar, default value is \code{spatstat.explore::fvnames(x, a = '.y')},
to speed up batch processes.}
}
\value{
Function \code{\link[=keyval.fv]{keyval.fv()}} returns a \link[base]{numeric} \link[base]{vector}.
}
\description{
Name and value of the \emph{black solid curve} as shown in \link[spatstat.explore]{plot.fv},
i.e., the primary outcome of an \link[spatstat.explore]{fv.object}.
}
\details{
Function \code{\link[=keyval.fv]{keyval.fv()}} finds the value of the (primary) outcome
of an \link[spatstat.explore]{fv.object}.
}
\keyword{internal}
