% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyval.R
\name{keyval}
\alias{keyval}
\alias{keyval.fv}
\title{Black-Solid-Curve in \link[spatstat.explore]{plot.fv}}
\usage{
keyval(x, ...)

\method{keyval}{fv}(x, key = fvnames(x, a = ".y"), .x = fvnames(x, a = ".x"), ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{additional parameters, currently of no use}

\item{key, .x}{\link[base]{character} scalars}
}
\value{
The \code{S3} method \code{\link[=keyval.fv]{keyval.fv()}} returns a \link[base]{numeric} \link[base]{vector}.
}
\description{
Name and value of the \emph{black solid curve} as shown in \link[spatstat.explore]{plot.fv},
i.e., the primary outcome of an \link[spatstat.explore]{fv.object}.
}
\details{
The \code{S3} method \code{\link[=keyval.fv]{keyval.fv()}} finds the value of the (primary) outcome
of an \link[spatstat.explore]{fv.object}.
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
