% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksmooth.fv.R
\name{ksmooth.fv}
\alias{ksmooth.fv}
\title{\link[stats]{ksmooth} of \link[spatstat.explore]{fv.object}}
\usage{
ksmooth.fv(x, key = fvnames(fv, a = ".y"), .x = fvnames(fv, a = ".x"), ...)
}
\arguments{
\item{x}{an \link[spatstat.explore]{fv.object}}

\item{key, .x}{\link[base]{character} scalars}

\item{...}{additional parameters of the functions
\link[stats]{ksmooth}, except for \code{x} and \code{y}}
}
\description{
\link[stats]{ksmooth} of \link[spatstat.explore]{fv.object}
}
\keyword{internal}
