% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastLegal.R
\name{lastLegal}
\alias{lastLegal}
\title{Last Legal Index}
\usage{
lastLegal(v)
}
\arguments{
\item{v}{\link[base]{double} \link[base]{vector}}
}
\description{
Last Legal Index
}
\details{
Legal, meaning not \code{0}, not \code{NaN} and not \code{Inf}.
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/nonS3/lastLegal.html}
}
\keyword{internal}
