% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperframe_S3.R
\name{length.hyperframe}
\alias{length.hyperframe}
\title{\code{S3} method dispatch of \link[base]{length} on \link[spatstat.geom]{hyperframe}}
\usage{
\method{length}{hyperframe}(x)
}
\arguments{
\item{x}{a \link[spatstat.geom]{hyperframe}}
}
\description{
\code{S3} method dispatch of \link[base]{length} on \link[spatstat.geom]{hyperframe}
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\seealso{
\link[spatstat.geom]{dim.hyperframe}
}
\keyword{internal}
