% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess.fv.R
\name{loess.fv}
\alias{loess.fv}
\title{\link[stats]{loess} of \link[spatstat.explore]{fv.object}}
\usage{
loess.fv(x, key = fvnames(fv, a = ".y"), .x = fvnames(fv, a = ".x"), ...)
}
\arguments{
\item{x}{an \link[spatstat.explore]{fv.object}}

\item{key, .x}{\link[base]{character} scalars}

\item{...}{additional parameters of the functions
\link[stats]{loess}, except for \code{formula} and \code{data}}
}
\description{
\link[stats]{loess} of \link[spatstat.explore]{fv.object}
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
