% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_S3.R
\name{log_ppp}
\alias{log_ppp}
\alias{log.ppp}
\alias{log1p.ppp}
\alias{log10.ppp}
\alias{log2.ppp}
\title{\link[base]{log} of \link[spatstat.geom]{ppp.object}}
\usage{
\method{log}{ppp}(x, base = exp(1))

\method{log1p}{ppp}(x)

\method{log10}{ppp}(x)

\method{log2}{ppp}(x)
}
\arguments{
\item{x}{a \link[spatstat.geom]{ppp.object}}

\item{base}{\link[base]{numeric} scalar}
}
\value{
Functions \code{\link[=log.ppp]{log.ppp()}}, \code{\link[=log1p.ppp]{log1p.ppp()}}, \code{\link[=log10.ppp]{log10.ppp()}} and \code{\link[=log2.ppp]{log2.ppp()}}
all return a \link[spatstat.geom]{ppp.object}.
}
\description{
...
}
\details{
Function \code{\link[=log.ppp]{log.ppp()}} takes a \link[base]{log} of continuous marks
of a \link[spatstat.geom]{ppp.object}.

Functions \code{\link[=log1p.ppp]{log1p.ppp()}}, \code{\link[=log10.ppp]{log10.ppp()}} and \code{\link[=log2.ppp]{log2.ppp()}} are similar.
}
\keyword{internal}
