% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_names.R
\name{mark_names}
\alias{mark_names}
\alias{mark_name<-}
\title{Mark Names}
\usage{
mark_names(x)

mark_name(x) <- value
}
\arguments{
\item{x}{\link[spatstat.geom]{ppp.object}}
}
\description{
Mark Names
}
\examples{
library(spatstat.data)
mark_names(betacells)

# ?waka
tryCatch(mark_names(waka), error = identity)
waka2 = waka
mark_name(waka2) = 'dbh' # tree diameter at breast height `dbh`
mark_names(waka2)
unstack(waka) # no name
unstack(waka2) # has name
}
\keyword{internal}
