% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{Math.fvlist}
\alias{Math.fvlist}
\title{\code{Math} \link[base]{groupGeneric} of \code{'fvlist'}}
\usage{
\method{Math}{fvlist}(x, ...)
}
\arguments{
\item{x}{an \code{'fvlist'}}

\item{...}{additional parameters for \code{Math} \link[base]{groupGeneric}}
}
\value{
Functions \code{\link[=Math.fvlist]{Math.fvlist()}} returns an \code{'fvlist'}.
}
\description{
...
}
\keyword{internal}
