% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_S3.R
\name{Math.ppp}
\alias{Math.ppp}
\title{\code{Math} \link[base]{groupGeneric} of \link[spatstat.geom]{ppp.object}}
\usage{
\method{Math}{ppp}(x, ...)
}
\arguments{
\item{x}{a \link[spatstat.geom]{ppp.object}}

\item{...}{additional parameters for \code{Math} \link[base]{groupGeneric}}
}
\value{
Functions \code{\link[=Math.ppp]{Math.ppp()}} returns a \link[spatstat.geom]{ppp.object}.
}
\description{
...
}
\details{
Function \code{\link[=Math.ppp]{Math.ppp()}} performs \code{Math} operations on \link[base]{numeric} \link[spatstat.geom]{marks}
of a \link[spatstat.geom]{ppp.object}.
}
\keyword{internal}
