% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{Math.ppplist}
\alias{Math.ppplist}
\title{\code{Math} \link[base]{groupGeneric} of \code{'ppplist'}}
\usage{
\method{Math}{ppplist}(x, ...)
}
\arguments{
\item{x}{an \code{'ppplist'}}

\item{...}{additional parameters for \code{Math} \link[base]{groupGeneric}}
}
\value{
Functions \code{\link[=Math.ppplist]{Math.ppplist()}} returns a \code{'ppplist'}.
}
\description{
...
}
\keyword{internal}
