% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{Math.psp}
\alias{Math.psp}
\title{\code{Math} \link[base]{groupGeneric} of \link[spatstat.geom]{psp.object}}
\usage{
\method{Math}{psp}(x, ...)
}
\arguments{
\item{x}{a \link[spatstat.geom]{psp.object}}

\item{...}{additional parameters for \code{Math} \link[base]{groupGeneric}}
}
\value{
Functions \code{\link[=Math.psp]{Math.psp()}} returns a \link[spatstat.geom]{psp.object}.
}
\description{
...
}
\details{
Function \code{\link[=Math.psp]{Math.psp()}} performs \code{Math} operations on the \link[base]{numeric} \link[spatstat.geom]{marks}
of a \link[spatstat.geom]{psp.object}.
}
\keyword{internal}
