% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{Math.tess}
\alias{Math.tess}
\title{\code{Math} \link[base]{groupGeneric} of \link[spatstat.geom]{tess}ellation}
\usage{
\method{Math}{tess}(x, ...)
}
\arguments{
\item{x}{a \link[spatstat.geom]{tess}ellation}

\item{...}{additional parameters for \code{Math} \link[base]{groupGeneric}}
}
\value{
Functions \code{\link[=Math.tess]{Math.tess()}} returns a \link[spatstat.geom]{ppp.object}.
}
\description{
...
}
\details{
Function \code{\link[=Math.tess]{Math.tess()}} performs \code{Math} operations on the \link[base]{numeric} \link[spatstat.geom]{marks}
of a \link[spatstat.geom]{tess}ellation.
}
\keyword{internal}
