% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_identical_by.R
\name{mc_identical_by}
\alias{mc_identical_by}
\title{Check if Data Column is Identical by Grouping}
\usage{
mc_identical_by(data, f, mc.cores = detectCores(), ...)
}
\arguments{
\item{data}{\link[base]{data.frame}}

\item{f}{\link[base]{factor}}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}.
Default is the return of the function \link[parallel]{detectCores}.}

\item{...}{additional parameters, currently not in use}
}
\description{
Check if Data Column is Identical by Grouping
}
\note{
The function \link[stats]{aggregate.data.frame} does not do parallel computing.

The function \code{collapse::collap} does not support \link[survival]{Surv} column.

Look more into \code{nlme:::collapse.groupedData}
}
\keyword{internal}
