% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette.R
\name{methods2kable}
\alias{methods2kable}
\title{\code{S3} methods table in vignette}
\usage{
methods2kable(
  generic.function,
  class,
  package,
  package_pattern,
  backtick = TRUE,
  ...
)
}
\arguments{
\item{generic.function, class}{see function \link[utils]{methods}}

\item{package}{\link[base]{character} scalar}

\item{package_pattern}{\link[base]{character} scalar of \link[base]{regex}}

\item{backtick}{\link[base]{logical} scalar, whether to put backticks around function names.
Default \code{TRUE} for Markdown/Quarto rendering.}

\item{...}{additional parameters of the function \link[utils]{methods}}
}
\value{
The function \code{\link[=methods2kable]{methods2kable()}} returns a \link[base]{data.frame}.
}
\description{
\code{S3} methods table in vignette
}
\keyword{internal}
