% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette.R
\name{Namespace2methodsInfo}
\alias{Namespace2methodsInfo}
\title{\code{\link[=Namespace2methodsInfo]{Namespace2methodsInfo()}}}
\usage{
Namespace2methodsInfo(ns, ...)
}
\arguments{
\item{ns}{\link[base]{character} scalar, package name, e.g., \code{'stats'}}

\item{...}{parameters of function \link[utils]{.S3methods}, other than \code{envir} and \code{useEnv}}
}
\description{
\code{\link[=Namespace2methodsInfo]{Namespace2methodsInfo()}}
}
\examples{
Namespace2methodsInfo(ns = 'stats', class = 'data.frame')

}
\keyword{internal}
