% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_hyperframe.R
\name{op_hyperframe}
\alias{op_hyperframe}
\title{Operations on \link[spatstat.geom]{hyperframe} with One-and-Only-One \link[spatstat.geom]{ppp}-\link[spatstat.geom:hyperframe]{hypercolumn}}
\usage{
op_hyperframe(X, ...)
}
\arguments{
\item{X}{a \link[spatstat.geom]{hyperframe}, containing \emph{\strong{one-and-only-one}} \link[spatstat.geom]{ppp}-\link[spatstat.geom:hyperframe]{hypercolumn}}

\item{...}{additional parameters of the function \code{\link[=op_ppplist]{op_ppplist()}}}
}
\value{
The function \code{\link[=op_hyperframe]{op_hyperframe()}} returns a \link[spatstat.geom]{hyperframe}.
}
\description{
Create hypercolumn(s) of
\link[spatstat.explore]{fv.object}s
or
distances
from
the one-and-only-one \link[spatstat.geom]{ppp}-\link[spatstat.geom:hyperframe]{hypercolumn}
inside a \link[spatstat.geom]{hyperframe}.
}
\keyword{internal}
