% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_ppplist.R
\name{op_ppplist}
\alias{op_ppplist}
\title{Batch Operations of \code{'ppplist'} Object}
\usage{
op_ppplist(x, op, mc.cores = detectCores(), ...)
}
\arguments{
\item{x}{a \code{'ppplist'} object}

\item{op}{workhorse \link[base]{function}, either \code{\link[=ppp_numeric2fv]{ppp_numeric2fv()}}, \code{\link[=ppp_multitype2fv]{ppp_multitype2fv()}} or \code{\link[=ppp2dist]{ppp2dist()}}}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}.
Default is the return of the function \link[parallel]{detectCores}.}

\item{...}{additional parameters of workhorse functions
\code{\link[=ppp_numeric2fv]{ppp_numeric2fv()}}, \code{\link[=ppp_multitype2fv]{ppp_multitype2fv()}} or \code{\link[=ppp2dist]{ppp2dist()}}}
}
\value{
The function \code{\link[=op_ppplist]{op_ppplist()}} returns a \link[stats]{listof}
\itemize{
\item function \code{\link[=ppp_numeric2fv]{ppp_numeric2fv()}} returns, if \code{op = ppp_numeric2fv}.
\item function \code{\link[=ppp_multitype2fv]{ppp_multitype2fv()}} returns, if \code{op = ppp_multitype2fv}.
\item function \code{\link[=ppp2dist]{ppp2dist()}} returns, if \code{op = ppp2dist}.
}
}
\description{
Batch operations for a \code{'ppplist'} input.
}
\details{
The function \code{\link[=op_ppplist]{op_ppplist()}} is a \pkg{parallel} batch process of
the workhorse function \code{\link[=ppp_numeric2fv]{ppp_numeric2fv()}}, \code{\link[=ppp_multitype2fv]{ppp_multitype2fv()}} or \code{\link[=ppp2dist]{ppp2dist()}}.
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
