% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_S3.R
\name{plot.pppkm}
\alias{plot.pppkm}
\title{Plot \code{'pppkm'} object}
\usage{
\method{plot}{pppkm}(
  x,
  main = "k-Means Clustering",
  ...,
  cols = pal_hue()(n = length(levels(f)))
)
}
\arguments{
\item{x}{a \code{'pppkm'} object, returned from function \code{\link[=kmeans.ppp]{kmeans.ppp()}}}

\item{main, ..., cols}{additional parameters of the function \link[spatstat.geom]{plot.ppp}}
}
\description{
Plot \code{'pppkm'} object
}
\keyword{internal}
