% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_ppp.R
\name{ppp2.}
\alias{ppp2.}
\alias{ppp2dist}
\alias{ppp_numeric2fv}
\alias{ppp_multitype2fv}
\title{Operations on \link[spatstat.geom]{ppp.object}}
\usage{
ppp2dist(x, fun, ...)

ppp_numeric2fv(x, fun, ...)

ppp_multitype2fv(x, fun, ...)
}
\arguments{
\item{x}{a \link[spatstat.geom]{ppp.object}}

\item{fun}{a distance \link[base]{function},
or a \link[base]{function} that returns an \link[spatstat.explore]{fv.object},
see \strong{Details}}

\item{...}{additional parameters of the function \code{fun}}
}
\value{
The functions \code{\link[=ppp_numeric2fv]{ppp_numeric2fv()}} and \code{\link[=ppp_multitype2fv]{ppp_multitype2fv()}} return a \link[stats]{listof}
\link[spatstat.explore]{fv.object}s.

The function \code{\link[=ppp2dist]{ppp2dist()}} returns a \link[stats]{listof}
\link[base]{double} \link[base]{vector}s.
}
\description{
Create
\link[spatstat.explore]{fv.object}s
or
distances
from a \link[spatstat.geom]{ppp.object}.
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
