% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pppBy.R
\name{pppBy}
\alias{pppBy}
\title{Hyper Data Frame with One-and-Only-One \link[spatstat.geom]{ppp}-Hyper Column}
\usage{
pppBy(
  marks,
  coords = ~x + y,
  by,
  data,
  window = owin(xrange = range(.x), yrange = range(.y)),
  ...
)
}
\arguments{
\item{marks}{one-sided \link[stats]{formula}, e.g.,
\code{~ m1+m2},
where \eqn{m_i}'s are one or more \link[spatstat.geom]{marks}}

\item{coords}{one-sided \link[stats]{formula}, variable names
of the \eqn{x}- and \eqn{y}-coordinates in \code{data}.
Default value is \code{~x+y}.}

\item{by}{two-sided \link[stats]{formula}}

\item{data}{\link[base]{data.frame}}

\item{window}{observation window \link[spatstat.geom]{owin},
default value is the \eqn{x}- and \eqn{y}-span of \code{coords} in \code{data}.}

\item{...}{additional parameters of the function \link[spatstat.geom]{ppp}}
}
\value{
The function \code{\link[=pppBy]{pppBy()}} returns a hyper data frame
with one-and-only-one
\link[spatstat.geom]{ppp}-hyper column.
}
\description{
To create a hyper data frame with one-and-only-one \link[spatstat.geom]{ppp}-hyper column.
}
\keyword{internal}
