% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtrapz.R
\name{print.cumv}
\alias{print.cumv}
\title{Print \code{'cumv'} Object}
\usage{
\method{print}{cumv}(x, ...)
}
\arguments{
\item{x}{a \code{'cumv'} object}

\item{...}{additional parameters, currently of no use}
}
\description{
Print \code{'cumv'} Object
}
\keyword{internal}
