% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_S3.R
\name{print.pppkm}
\alias{print.pppkm}
\title{Print \code{'pppkm'} object}
\usage{
\method{print}{pppkm}(x, ...)
}
\arguments{
\item{x}{a \code{'pppkm'} object, returned from function \code{\link[=kmeans.ppp]{kmeans.ppp()}}}

\item{...}{additional parameters, currently no use}
}
\description{
Print \code{'pppkm'} object
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
