% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorlist.R
\name{print.vectorlist}
\alias{print.vectorlist}
\title{Print a \code{'vectorlist'}}
\usage{
\method{print}{vectorlist}(x, ...)
}
\arguments{
\item{x}{a \code{'vectorlist'}}

\item{...}{additional parameters, currently of no use}
}
\description{
Print a \code{'vectorlist'}
}
\keyword{internal}
