% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{quantile.anylist}
\alias{quantile.anylist}
\title{Quantiles of \link[base]{numeric} \link[spatstat.geom]{marks} in \link[spatstat.geom]{anylist}}
\usage{
\method{quantile}{anylist}(x, ...)
}
\arguments{
\item{x}{an \link[spatstat.geom]{anylist}}

\item{...}{additional parameters of the function \code{stats:::quantile.default()}}
}
\description{
Quantiles of \link[base]{numeric} \link[spatstat.geom]{marks} in \link[spatstat.geom]{anylist}
}
\keyword{internal}
