% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{quantile.hyperframe}
\alias{quantile.hyperframe}
\title{Quantiles of \link[base]{numeric} elements in \link[spatstat.geom]{hyperframe}}
\usage{
\method{quantile}{hyperframe}(x, ...)
}
\arguments{
\item{x}{a \link[spatstat.geom]{hyperframe}}

\item{...}{additional parameters of the function \code{stats:::quantile.default()}}
}
\description{
Quantiles of \link[base]{numeric} elements in \link[spatstat.geom]{hyperframe}
}
\keyword{internal}
