% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{quantile.ppp}
\alias{quantile.ppp}
\title{Quantiles of \link[base]{numeric} \link[spatstat.geom]{marks} in \link[spatstat.geom]{ppp.object}}
\usage{
\method{quantile}{ppp}(x, ...)
}
\arguments{
\item{x}{a \link[spatstat.geom]{ppp.object}}

\item{...}{additional parameters of the function \code{stats:::quantile.default()}}
}
\description{
Quantiles of \link[base]{numeric} \link[spatstat.geom]{marks} in \link[spatstat.geom]{ppp.object}
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
