% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{quantile.ppplist}
\alias{quantile.ppplist}
\title{Quantiles of \link[base]{numeric} \link[spatstat.geom]{marks} in \code{'ppplist'}}
\usage{
\method{quantile}{ppplist}(x, ...)
}
\arguments{
\item{x}{a \code{'ppplist'}}

\item{...}{additional parameters of the function \code{stats:::quantile.default()}}
}
\description{
Quantiles of \link[base]{numeric} \link[spatstat.geom]{marks} in \code{'ppplist'}
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
