% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette.R
\name{rds2versiondate}
\alias{rds2versiondate}
\title{\link{rds2versiondate}}
\usage{
rds2versiondate(x, pkg, format = "\%a \%b \%d, \%Y", ...)
}
\arguments{
\item{x}{\link[base]{matrix}, e.g., from \url{https://cran.r-project.org/web/packages/packages.rds}}

\item{pkg}{\link[base]{character} \link[base]{vector}}

\item{format}{\link[base]{Date} format, see functions \link[base]{format.Date} and \link[base]{strptime} for detail}

\item{...}{additional parameters, currently of no use}
}
\description{
\link{rds2versiondate}
}
\keyword{internal}
