% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlabelEnv_test.R
\name{rlabelEnv_test}
\alias{rlabelEnv_test}
\alias{rlabelEnv.test}
\alias{rlabelEnv.test.ppp}
\alias{rlabelEnv.test.ppplist}
\alias{rlabelEnv.test.hyperframe}
\title{Random Re-Labelling Envelope Test}
\usage{
rlabelEnv.test(x, ...)

\method{rlabelEnv.test}{ppp}(x, ...)

\method{rlabelEnv.test}{ppplist}(x, ...)

\method{rlabelEnv.test}{hyperframe}(x, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{arguments of function \link[spatstat.explore]{envelope.ppp},
other than parameters \code{simulate},
\code{savefuns} and \code{verbose}}
}
\description{
Random Re-Labelling Envelope Test
}
\examples{
set.seed(52); re1_anemones = spatstat.data::anemones |>
  rlabelEnv.test(fun = spatstat.explore::Kmark)
re1_anemones |> class()
re1_anemones |> attr(which = 'p', exact = TRUE)

set.seed(52); re1_anemones_times = spatstat.data::anemones |>
  rlabelEnv.test(fun = spatstat.explore::Kmark, f = `*`)
re1_anemones_times |> attr(which = 'p', exact = TRUE)

stopifnot(!identical(re1_anemones, re1_anemones_times))
  
\donttest{
set.seed(52); spatstat.data::anemones |>
  rlabelEnv.test(fun = spatstat.explore::Kmark, f = \(m1, m2) { m1*m2 }) |>
  identical(y = re1_anemones_times) |>
  stopifnot()
}

set.seed(12); re_ants = spatstat.data::ants |>
  rlabelEnv.test(fun = spatstat.explore::Gcross)
re_ants |> attr(which = 'p', exact = TRUE)

}
\keyword{internal}
