% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.groupedHyperframe.R
\name{setGroups}
\alias{setGroups}
\alias{setGroups.hyperframe}
\title{Creation of \code{groupedHyperframe}}
\usage{
setGroups(x, group, ...)

\method{setGroups}{hyperframe}(x, group, ...)
}
\arguments{
\item{x}{see Usage}

\item{group}{\link[stats]{formula}}

\item{...}{additional parameters}
}
\value{
The \code{S3} generic function \code{\link[=setGroups]{setGroups()}} returns an object of derived class \code{'grouped*'}.
}
\description{
To create a \code{groupedHyperframe} object.
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
