% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.spline.fv.R
\name{smooth.spline.fv}
\alias{smooth.spline.fv}
\title{\link[stats]{smooth.spline} of \link[spatstat.explore]{fv.object}}
\usage{
smooth.spline.fv(
  x,
  key = fvnames(fv, a = ".y"),
  .x = fvnames(fv, a = ".x"),
  ...
)
}
\arguments{
\item{x}{an \link[spatstat.explore]{fv.object}}

\item{key, .x}{\link[base]{character} scalars}

\item{...}{additional parameters of the functions
\link[stats]{smooth.spline}, except for \code{x}, \code{y} and \code{keep.data}}
}
\description{
\link[stats]{smooth.spline} of \link[spatstat.explore]{fv.object}
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
