% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_S3.R
\name{split.hyperframekm}
\alias{split.hyperframekm}
\title{\code{\link[=split.hyperframekm]{split.hyperframekm()}}}
\usage{
\method{split}{hyperframekm}(x, ...)
}
\arguments{
\item{x}{a \code{'hyperframekm'}, returned from function \code{\link[=kmeans.hyperframe]{kmeans.hyperframe()}}}

\item{...}{additional parameters, currently no use}
}
\value{
The function \code{\link[=split.hyperframekm]{split.hyperframekm()}} returns a \code{'groupedHyperframe'}.
}
\description{
\code{\link[=split.hyperframekm]{split.hyperframekm()}}
}
\keyword{internal}
