% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_S3.R
\name{split.pppkmlist}
\alias{split.pppkmlist}
\title{\code{split.pppkmlist}}
\usage{
\method{split}{pppkmlist}(x, ...)
}
\arguments{
\item{x}{a \code{'pppkmlist'} object, returned from function \code{\link[=kmeans.ppplist]{kmeans.ppplist()}}}

\item{...}{additional parameters, currently no use}
}
\value{
The function \code{\link[=split.pppkmlist]{split.pppkmlist()}} returns a \code{'splitppp'} object from the
workhorse function \link[spatstat.geom]{split.ppp}.
}
\description{
\code{split.pppkmlist}
}
\keyword{internal}
