% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_kmeans.R
\name{split_kmeans}
\alias{split_kmeans}
\alias{split_kmeans.default}
\alias{split_kmeans.listof}
\alias{split_kmeans.hyperframe}
\title{\link[base]{split} by \link[stats]{kmeans} Clustering}
\usage{
split_kmeans(x, ...)

\method{split_kmeans}{default}(x, ...)

\method{split_kmeans}{listof}(x, ...)

\method{split_kmeans}{hyperframe}(x, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{additional parameters of function \link[stats]{kmeans}}
}
\description{
\link[base]{split} by \link[stats]{kmeans} clustering
}
\note{
Function \code{\link[=split_kmeans.default]{split_kmeans.default()}} is supposed to work with
\link[spatstat.geom]{ppp.object}.
}
\examples{
library(spatstat.data)
library(spatstat.geom)
shapley |> plot()
r = shapley |> split_kmeans(formula = ~ x + y + Mag, centers = 3L)
r[[1L]] |> plot()
r[[2L]] |> plot()
r[[3L]] |> plot()
library(spatstat.data)
library(spatstat.geom)
flu$pattern[] = flu$pattern |> 
 lapply(FUN = `mark_name<-`, value = 'stain') # read ?flu carefully
 
# split_kmeans.listof()  
flu$pattern[1:2] |> split_kmeans(formula = ~ x + y, centers = 3L) 

# split_kmeans.hyperframe() (accepts 'groupedHyperframe'!!)
flu[1:2,] |> split_kmeans(formula = ~ x + y, centers = 3L)

}
\keyword{internal}
