% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtrapz.R
\name{[.cumv}
\alias{[.cumv}
\title{Extract Parts of \code{'cumv'} Object}
\usage{
\method{[}{cumv}(x, i)
}
\arguments{
\item{x}{a \code{'cumv'} object}

\item{i}{\link[base]{integer} scalar or \link[base]{vector}; or \link[base]{logical} \link[base]{vector}}
}
\description{
Extract Parts of \code{'cumv'} Object
}
\keyword{internal}
