% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedHyperframe.R
\name{[.groupedHyperframe}
\alias{[.groupedHyperframe}
\title{Extract Subset of \link{groupedHyperframe}}
\usage{
\method{[}{groupedHyperframe}(x, ...)
}
\arguments{
\item{x}{a \link{groupedHyperframe}}

\item{...}{additional parameters of \link[spatstat.geom]{[.hyperframe}}
}
\value{
The function \link{[.groupedHyperframe} returns a \link{groupedHyperframe} or a \link[spatstat.geom]{hyperframe}.
}
\description{
Extract Subset of \link{groupedHyperframe}
}
\keyword{internal}
