% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_ppp_tzh.R
\name{[.ppp_tzh}
\alias{[.ppp_tzh}
\title{\verb{[.ppp_tzh}}
\usage{
\method{[}{ppp_tzh}(x, i, j, drop = FALSE, ..., clip = FALSE)
}
\arguments{
\item{x}{an object of internal class \code{'ppp_tzh'},
a derived class from \link[spatstat.geom]{ppp.object}}

\item{i, j}{see function \link[spatstat.geom]{[.ppp}}

\item{drop}{..}

\item{...}{..}

\item{clip}{..}
}
\value{
The function \code{`[.ppp_tzh`} returns a \link[spatstat.geom]{ppp.object},
\strong{not} an object of class \code{'ppp_tzh'}.
}
\description{
\verb{[.ppp_tzh}
}
\keyword{internal}
