% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvlist.R
\name{summary.fvlist}
\alias{summary.fvlist}
\title{Truncated Summary of \code{'fvlist'}}
\usage{
\method{summary}{fvlist}(object, mc.cores = getOption("cores"), ...)
}
\arguments{
\item{object}{an \code{'fvlist'}}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}.
Default is the return of function \link[parallel]{detectCores}.}

\item{...}{additional parameters of (internal) function \code{trunc_id.fvlist()}}
}
\description{
Truncated Summary of \code{'fvlist'}
}
\keyword{internal}
