% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedHyperframe.R
\name{summary.groupedHyperframe}
\alias{summary.groupedHyperframe}
\title{Summary Information of \link{groupedHyperframe}}
\usage{
\method{summary}{groupedHyperframe}(object, ...)
}
\arguments{
\item{object}{a \link{groupedHyperframe}}

\item{...}{additional parameters, currently not in use}
}
\value{
The function \code{\link[=summary.groupedHyperframe]{summary.groupedHyperframe()}} returns an R object of class \code{'summary.groupedHyperframe'}.
}
\description{
Summary Information of \link{groupedHyperframe}
}
\keyword{internal}
