% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_fv.R
\name{summary_fv}
\alias{summary_fv}
\title{Summarize \link[spatstat.explore]{fv}-\code{hypercolumns}}
\usage{
summary_fv(X, rmax, ...)
}
\arguments{
\item{X}{a \link[spatstat.geom]{hyperframe} or \link{groupedHyperframe},
containing one or more \link[spatstat.explore]{fv.object} column(s)}

\item{rmax}{\link[base]{numeric} scalar, user-specified truncation point \eqn{r_\text{max}}}

\item{...}{additional parameters of function \code{\link[=summary.fvlist]{summary.fvlist()}}, most importantly \code{mc.cores}.}
}
\value{
Function \code{\link[=summary_fv]{summary_fv()}} returns a \link[spatstat.geom]{hyperframe} or \link{groupedHyperframe}.
}
\description{
Summarize \link[spatstat.explore]{fv}-\code{hypercolumns}
}
\keyword{internal}
