% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superimpose.hyperframe.R
\name{superimpose.groupedHyperframe}
\alias{superimpose.groupedHyperframe}
\title{superimpose.groupedHyperframe}
\usage{
\method{superimpose}{groupedHyperframe}(...)
}
\arguments{
\item{...}{one or more \link{groupedHyperframe}s}
}
\value{
Function \code{\link[=superimpose.groupedHyperframe]{superimpose.groupedHyperframe()}} returns a \link{groupedHyperframe}.
}
\description{
superimpose.groupedHyperframe
}
\keyword{internal}
