% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorlist.R
\name{t.vectorlist}
\alias{t.vectorlist}
\title{Transpose a \code{'vectorlist'}}
\usage{
\method{t}{vectorlist}(x)
}
\arguments{
\item{x}{a \code{'vectorlist'}}
}
\value{
The \code{S3} method dispatch \code{\link[=t.vectorlist]{t.vectorlist()}} returns
a \code{'vectorlist'} of equi-\link[base]{length}.
}
\description{
Transpose a \code{'vectorlist'}
}
\details{
tzh defines a derived class \code{'vectorlist'},
i.e., a \link[stats]{listof} \link[base]{vector}s,
which \link[base]{inherits} from
\link[spatstat.geom]{anylist}.
The implementation of \code{'vectorlist'} is
inspired by class \link[spatstat.geom]{solist}.

The \code{S3} method dispatch \code{\link[=t.vectorlist]{t.vectorlist()}},
of the generic function \link[base]{t},
transposes a \code{'vectorlist'} of equi-\link[base]{length}.
We illustrate this concept using data set
\link[spatstat.data]{Kovesi} in \strong{Examples}.
}
\note{
The motivation of
the derived class \code{'vectorlist'} and
the method dispatch \code{\link[=t.vectorlist]{t.vectorlist()}}
is that
The function \link[spatstat.geom]{with.hyperframe}
could be slow in a batch process.
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
