% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapz.R
\name{trapz}
\alias{trapz}
\alias{trapz.fv}
\alias{cumtrapz.fv}
\alias{cumtrapz.fvlist}
\alias{cumtrapz.hyperframe}
\title{Trapzoidal Integration of \link[spatstat.explore]{fv.object}}
\usage{
trapz.fv(x, key = fvnames(x, a = ".y"), .x = fvnames(x, a = ".x"), ...)

cumtrapz.fv(x, key = fvnames(x, a = ".y"), .x = fvnames(x, a = ".x"), ...)

cumtrapz.fvlist(x, mc.cores = getOption("cores"), ...)

cumtrapz.hyperframe(x, rmax, ...)
}
\arguments{
\item{x}{an \link[spatstat.explore]{fv.object}}

\item{key, .x}{\link[base]{character} scalars}

\item{...}{additional parameters}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}.
Default is the return of function \link[parallel]{detectCores}.}

\item{rmax}{\link[base]{numeric} scalar, user-specified truncation point \eqn{r_\text{max}}}
}
\value{
Functions \code{\link[=trapz.fv]{trapz.fv()}} and \code{\link[=vtrapz.fv]{vtrapz.fv()}} return a \link[base]{numeric} scalar.

Functions \code{\link[=cumtrapz.fv]{cumtrapz.fv()}} and \code{\link[=cumvtrapz.fv]{cumvtrapz.fv()}} return a \link[base]{numeric} \link[base]{vector}.
}
\description{
Trapzoidal Integration of \link[spatstat.explore]{fv.object}
}
\details{
Functions \code{\link[=trapz.fv]{trapz.fv()}} and \code{\link[=cumtrapz.fv]{cumtrapz.fv()}}
obtain the (cumulative)
\link[pracma]{trapz}oidal integration of the area under the primary outcome
of a function value \link[spatstat.explore]{fv.object}.
}
\keyword{internal}
