% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapz.R
\name{trapz_fv}
\alias{trapz_fv}
\alias{trapz.fv}
\alias{cumtrapz.fv}
\title{Trapzoidal Integration of \link[spatstat.explore]{fv.object}}
\usage{
trapz.fv(x, key = fvnames(x, a = ".y"))

cumtrapz.fv(x, key = fvnames(x, a = ".y"))
}
\arguments{
\item{x}{an \link[spatstat.explore]{fv.object}}

\item{key}{\link[base]{character} scalar, see function \code{\link[=keyval.fv]{keyval.fv()}}}
}
\value{
Functions \code{\link[=trapz.fv]{trapz.fv()}} and \code{\link[=vtrapz.fv]{vtrapz.fv()}} return a \link[base]{numeric} scalar.

Functions \code{\link[=cumtrapz.fv]{cumtrapz.fv()}} and \code{\link[=cumvtrapz.fv]{cumvtrapz.fv()}} return a \link[base]{numeric} \link[base]{vector}.
}
\description{
Trapzoidal Integration of \link[spatstat.explore]{fv.object}
}
\details{
Functions \code{\link[=trapz.fv]{trapz.fv()}} and \code{\link[=cumtrapz.fv]{cumtrapz.fv()}}
obtain the (cumulative)
\link[pracma]{trapz}oidal integration of the area under the primary outcome
of a function value \link[spatstat.explore]{fv.object}.
}
\keyword{internal}
