% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmark_ext.R
\name{unmark.ppplist}
\alias{unmark.ppplist}
\title{\code{\link[=unmark.ppplist]{unmark.ppplist()}}}
\usage{
\method{unmark}{ppplist}(X)
}
\arguments{
\item{X}{a \code{'ppplist'}}
}
\description{
\code{\link[=unmark.ppplist]{unmark.ppplist()}}
}
\details{
Function \code{\link[=unmark.ppplist]{unmark.ppplist()}} is a iteration of \link[spatstat.geom]{unmark.ppp},
but not as yet \link[spatstat.geom]{unmark.splitppp}.
}
\keyword{internal}
