% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmark_ext.R
\name{unmark_}
\alias{unmark_}
\alias{unmark_.ppplist}
\title{Batch Process of \link[spatstat.geom]{unmark}}
\usage{
unmark_(X)

\method{unmark_}{ppplist}(X)
}
\arguments{
\item{X}{see \strong{Usage}}
}
\description{
Batch Process of \link[spatstat.geom]{unmark}
}
\details{
Function \code{\link[=unmark_.ppplist]{unmark_.ppplist()}} is a iteration of \link[spatstat.geom]{unmark.ppp}.

Note that Dr. Baddeley has told me that \link[spatstat.geom]{unmark.splitppp} is a bandage fix.
}
\keyword{internal}
