% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette.R
\name{vignette_methods}
\alias{vignette_methods}
\title{\code{S3} methods table in vignette}
\usage{
vignette_methods(class, package, ...)
}
\arguments{
\item{class}{\link[base]{character} scalar}

\item{package}{\link[base]{character} scalar}

\item{...}{additional parameters of function \link[utils]{methods}}
}
\value{
Function \code{\link[=vignette_methods]{vignette_methods()}} returns a \link[base]{data.frame}.
}
\description{
\code{S3} methods table in vignette
}
\keyword{internal}
