% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtrapz.R
\name{vtrapz}
\alias{vtrapz}
\title{Average Vertical Height of Trapezoidal Integration}
\usage{
vtrapz(x, ...)
}
\arguments{
\item{x}{\link[base]{numeric} \link[base]{vector}}

\item{...}{additional parameters of the function \link[pracma]{trapz}}
}
\value{
The function \code{\link[=vtrapz]{vtrapz()}} return a \link[base]{numeric} scalar.
}
\description{
Trapezoidal integration divided by \eqn{x}-domain.
}
\note{
This is a tentative thought: the prefix \code{v} stands for 'vertical'.
}
\keyword{internal}
