% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{wrobel_lung}
\alias{wrobel_lung}
\title{wrobel_lung}
\format{
A \link[base]{data.frame}
}
\usage{
wrobel_lung
}
\description{
From Dr. Julia Wrobel's data \url{https://sph.emory.edu/profile/faculty/julia-wrobel}.
}
\keyword{internal}
