% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.groupedHyperframe}
\alias{aggregate.groupedHyperframe}
\title{Aggregate \link{groupedHyperframe}}
\usage{
\method{aggregate}{groupedHyperframe}(x, by, ...)
}
\arguments{
\item{x}{a \link{groupedHyperframe}}

\item{by}{a one-sided \link[stats]{formula}}

\item{...}{additional parameters of function \code{\link[=aggregate.vectorlist]{aggregate.vectorlist()}},
most importantly parameter \code{fun}}
}
\value{
Function \code{\link[=aggregate.groupedHyperframe]{aggregate.groupedHyperframe()}} returns a \link[spatstat.geom]{hyperframe}.
}
\description{
Aggregate \link{groupedHyperframe}
}
\keyword{internal}
