% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.hyperframe.R
\name{aggregate.hyperframe}
\alias{aggregate.hyperframe}
\title{Aggregate Hyper Data Frame}
\usage{
\method{aggregate}{hyperframe}(x, by, ...)
}
\arguments{
\item{x}{a \link[spatstat.geom]{hyperframe}}

\item{by}{a one-sided \link[stats]{formula},
containing regular-column names of the input \code{x}}

\item{...}{additional parameters, currently not in use}
}
\value{
The \code{S3} method \code{\link[=aggregate.hyperframe]{aggregate.hyperframe()}} returns a \link[spatstat.geom]{hyperframe}.
}
\description{
To \link[stats]{aggregate} a \link[spatstat.geom]{hyperframe}.
}
