% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.hyperframe}
\alias{aggregate.hyperframe}
\title{Aggregate Hyper Data Frame}
\usage{
\method{aggregate}{hyperframe}(x, by, ...)
}
\arguments{
\item{x}{a \link[spatstat.geom]{hyperframe}}

\item{by}{a one-sided \link[stats]{formula},
containing regular-column names of the input \code{x}}

\item{...}{additional parameters of the function \code{\link[=aggregate.vectorlist]{aggregate.vectorlist()}},
most importantly the parameter \code{fun}}
}
\value{
The \code{S3} method \code{\link[=aggregate.hyperframe]{aggregate.hyperframe()}} returns a \link[spatstat.geom]{hyperframe}.
}
\description{
To aggregate the aggregatable hypercolumns of a
\link[spatstat.geom]{hyperframe}.
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
