% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_marks.R
\name{aggregate_marks}
\alias{aggregate_marks}
\alias{aggregate_marks.ppp}
\alias{aggregate_marks.ppplist}
\title{\link{aggregate_marks}}
\usage{
aggregate_marks(x, ...)

\method{aggregate_marks}{ppp}(x, ...)

\method{aggregate_marks}{ppplist}(x, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{additional parameters of function \link[stats]{aggregate.data.frame}}
}
\description{
\link{aggregate_marks}
}
\keyword{internal}
