% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvlist.R
\name{as.fvlist}
\alias{as.fvlist}
\title{Convert a \link[stats]{listof} \link[spatstat.explore]{fv.object}s into \code{'fvlist'}}
\usage{
as.fvlist(X, data.name)
}
\arguments{
\item{X}{a \link[stats]{listof} \link[spatstat.explore]{fv.object}s}

\item{data.name}{\link[base]{character} scalar, name of \code{X}, for console message output}
}
\value{
Function \code{\link[=as.fvlist]{as.fvlist()}} returns an \link[base]{invisible} \link[base]{list} with elements
\describe{
\item{\verb{$r}}{\link[base]{numeric} \link[base]{vector}, the \eqn{r}-values}
\item{\verb{$.y}}{\link[base]{character} scalar, the return of \code{spatstat.explore::fvnames(x, a = '.y')}}
\item{\verb{$rmax}}{\link[base]{numeric} scalar or \link[base]{vector}, the \link[base]{unique} values of the legal \eqn{r_\text{max}} of each \link[spatstat.explore]{fv.object}, if any one of them is less than the user-specified \eqn{r_\text{max}}}
}

!!! 'Legal' not documented yet!!!
}
\description{
Convert a \link[stats]{listof} \link[spatstat.explore]{fv.object}s into \code{'fvlist'}
}
\keyword{internal}
