% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.groupedHyperframe.R
\name{as.groupedHyperframe}
\alias{as.groupedHyperframe}
\alias{as.groupedHyperframe.hyperframe}
\alias{as.groupedHyperframe.data.frame}
\title{Creation of \code{groupedHyperframe}}
\usage{
as.groupedHyperframe(x, group, ...)

\method{as.groupedHyperframe}{hyperframe}(x, group, ...)

\method{as.groupedHyperframe}{data.frame}(x, group, ...)
}
\arguments{
\item{x}{see Usage}

\item{group}{\link[stats]{formula}}

\item{...}{additional parameters}
}
\value{
Function \code{\link[=as.groupedHyperframe]{as.groupedHyperframe()}} returns a \code{groupedHyperframe}.
}
\description{
To create a \code{groupedHyperframe} object
}
\keyword{internal}
