% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorlist.R
\name{as.vectorlist}
\alias{as.vectorlist}
\title{\link{as.vectorlist}}
\usage{
as.vectorlist(x, ...)
}
\arguments{
\item{x}{an \link[spatstat.geom]{anylist}}

\item{...}{additional parameters of the function \code{\link[=is.vectorlist]{is.vectorlist()}}}
}
\description{
\link{as.vectorlist}
}
\keyword{internal}
