% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aug4gam.R
\name{aug4gam}
\alias{aug4gam}
\alias{aug4gam.data.frame}
\alias{aug4gam.hyperframe}
\alias{aug4gam.vectorlist}
\title{Augment Hypercolumn(s) for \link[mgcv]{gam}}
\usage{
aug4gam(x, ...)

\method{aug4gam}{data.frame}(x, ...)

\method{aug4gam}{hyperframe}(x, ...)

\method{aug4gam}{vectorlist}(x, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{additional parameters, currently of no use}
}
\description{
Augment \emph{all} \link[base]{numeric} \link[base]{vector} hypercolumns in a \link[spatstat.geom]{hyperframe}
for \link[mgcv]{gam}.
}
\author{
Tingting Zhan, Erjia Cui
}
