% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_hyperframe.R
\name{batch}
\alias{batch}
\alias{Emark_}
\alias{Emark_.hyperframe}
\alias{Emark_.ppplist}
\alias{Vmark_}
\alias{Vmark_.hyperframe}
\alias{Vmark_.ppplist}
\alias{Kmark_}
\alias{Kmark_.hyperframe}
\alias{Kmark_.ppplist}
\alias{markcorr_}
\alias{markcorr_.hyperframe}
\alias{markcorr_.ppplist}
\alias{markvario_}
\alias{markvario_.hyperframe}
\alias{markvario_.ppplist}
\alias{Gcross_}
\alias{Gcross_.hyperframe}
\alias{Gcross_.ppplist}
\alias{Jcross_}
\alias{Jcross_.hyperframe}
\alias{Jcross_.ppplist}
\alias{Kcross_}
\alias{Kcross_.hyperframe}
\alias{Kcross_.ppplist}
\alias{Lcross_}
\alias{Lcross_.hyperframe}
\alias{Lcross_.ppplist}
\alias{markconnect_}
\alias{markconnect_.hyperframe}
\alias{markconnect_.ppplist}
\alias{nncross_}
\alias{nncross_.hyperframe}
\alias{nncross_.ppplist}
\title{Batch Process}
\usage{
Emark_(X, ...)

\method{Emark_}{hyperframe}(X, correction = "none", ...)

\method{Emark_}{ppplist}(X, correction = "none", ...)

Vmark_(X, ...)

\method{Vmark_}{hyperframe}(X, correction = "none", ...)

\method{Vmark_}{ppplist}(X, correction = "none", ...)

Kmark_(X, ...)

\method{Kmark_}{hyperframe}(X, correction = "none", ...)

\method{Kmark_}{ppplist}(X, correction = "none", ...)

markcorr_(X, ...)

\method{markcorr_}{hyperframe}(X, correction = "none", ...)

\method{markcorr_}{ppplist}(X, correction = "none", ...)

markvario_(X, ...)

\method{markvario_}{hyperframe}(X, correction = "none", ...)

\method{markvario_}{ppplist}(X, correction = "none", ...)

Gcross_(X, ...)

\method{Gcross_}{hyperframe}(X, correction = "none", ...)

\method{Gcross_}{ppplist}(X, correction = "none", ...)

Jcross_(X, ...)

\method{Jcross_}{hyperframe}(X, correction = "none", ...)

\method{Jcross_}{ppplist}(X, correction = "none", ...)

Kcross_(X, ...)

\method{Kcross_}{hyperframe}(X, correction = "none", ...)

\method{Kcross_}{ppplist}(X, correction = "none", ...)

Lcross_(X, ...)

\method{Lcross_}{hyperframe}(X, correction = "none", ...)

\method{Lcross_}{ppplist}(X, correction = "none", ...)

markconnect_(X, ...)

\method{markconnect_}{hyperframe}(X, correction = "none", ...)

\method{markconnect_}{ppplist}(X, correction = "none", ...)

nncross_(X, ...)

\method{nncross_}{hyperframe}(X, ...)

\method{nncross_}{ppplist}(X, ...)
}
\arguments{
\item{X}{a \link[spatstat.geom]{hyperframe}}

\item{...}{additional parameters of user operation}

\item{correction}{\link[base]{character} scalar,
see functions
\link[spatstat.explore]{markcorr},
\link[spatstat.explore]{Gcross},
etc.
Default \code{'none'} to save computing time.}
}
\description{
See workhorse function \code{\link[=op_hyperframe]{op_hyperframe()}}.
}
\details{
User Interface of Operations on \link[spatstat.geom]{hyperframe} with One-and-Only-One \link[spatstat.geom]{ppp}-\link[spatstat.geom:hyperframe]{hypercolumn}
}
\examples{
# in \CRANpkg{spatstat.data}
# no good example for [Emark_]
# no hyperframe with ppp-hypercolumn with numeric marks

library(spatstat.geom)
fluM = spatstat.data::flu |>
 subset(subset = (stain == 'M2-M1') & (virustype == 'wt'))
fluM
r = seq.int(from = 0, to = 100, by = 5)
fluM |>
 Gcross_(i = 'M1', j = 'M2', r = r)
fluM |>
 nncross_(i = 'M1', j = 'M2')
}
\keyword{internal}
