% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{bib}
\alias{bib}
\alias{bioinformatics_btaf430}
\title{\link[utils]{bibentry} for Quarto Book}
\format{
An object of class \code{bibentry} of length 1.
}
\usage{
bioinformatics_btaf430
}
\description{
\link[utils]{bibentry} for Quarto Book
}
\keyword{internal}
