% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cash_set.R
\name{$<-.groupedHyperframe}
\alias{$<-.groupedHyperframe}
\title{Syntactic Sugar for \link{groupedHyperframe} Object}
\usage{
\method{$}{groupedHyperframe}(x, name) <- value
}
\arguments{
\item{x}{a groupedHyperframe}

\item{name, value}{see function \link[spatstat.geom]{$<-.hyperframe}}
}
\value{
Syntactic sugar \code{\link[=$<-.groupedHyperframe]{$<-.groupedHyperframe()}} returns a \link{groupedHyperframe}.
}
\description{
Syntactic Sugar for \link{groupedHyperframe} Object
}
\note{
Function \link[spatstat.geom]{$<-.hyperframe} drops \link[base]{attributes}.
}
\keyword{internal}
