% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fvlist.R
\name{check_fvlist}
\alias{check_fvlist}
\title{check_fvlist}
\usage{
check_fvlist(X, data.name = deparse1(substitute(X)))
}
\arguments{
\item{X}{a \link[stats]{listof} \link[spatstat.explore]{fv.object}s}

\item{data.name}{\link[base]{character} scalar, name of \code{X}, for console message output}
}
\value{
Function \code{\link[=check_fvlist]{check_fvlist()}} returns an \link[base]{invisible} \link[base]{list} with elements
\describe{
\item{\verb{$r}}{\link[base]{numeric} \link[base]{vector}, the \eqn{r}-values}
\item{\verb{$.y}}{\link[base]{character} scalar, the return of \code{spatstat.explore::fvnames(x, a = '.y')}}
\item{\verb{$rmax}}{\link[base]{numeric} scalar or \link[base]{vector}, the \link[base]{unique} values of the legal \eqn{r_\text{max}} of each \link[spatstat.explore]{fv.object}, if any one of them is less than the user-specified \eqn{r_\text{max}}}
}
}
\description{
A helper function to check the validity of a \link[stats]{listof} \link[spatstat.explore]{fv.object}s.
}
\details{
Function \code{\link[=check_fvlist]{check_fvlist()}} checks that whether all \link[spatstat.explore]{fv.object}s
in the input has the same
\itemize{
\item {\eqn{x}-axis, or the \eqn{r}-values}
\item {\code{attr(,'fname')}, see explanation of this \link[base]{attributes} in function \link[spatstat.explore]{fv}}
\item {\code{spatstat.explore::fvnames(x, a = '.y')} returns}
}

!!! 'Legal' not documented yet!!!
}
\keyword{internal}
