% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Defunct.R
\name{cli_}
\alias{cli_}
\alias{cli_RPubs_}
\alias{cli_doi_}
\title{\link[base]{.Defunct} Messages using Package \CRANpkg{cli}}
\usage{
cli_RPubs_(x)

cli_doi_(x)
}
\arguments{
\item{x}{\link[base]{character} scalar}
}
\description{
Internal helper functions,
to display beautiful \link[base]{.Defunct} messages using package \CRANpkg{cli}.
}
\examples{
cli_RPubs_('tingtingzhan/groupedHyperframe')
cli_doi_('10.1002/bimj.4710230408')

}
\keyword{internal}
