% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Defunct.R
\name{defunct}
\alias{defunct}
\alias{aggregate_quantile}
\title{Defunct Functions}
\usage{
aggregate_quantile(...)
}
\arguments{
\item{...}{Defunct parameters}
}
\description{
The functions mentioned in hard-copy journals, but later \link[base]{.Defunct}.
}
\keyword{internal}
