% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Defunct.R
\name{defunct}
\alias{defunct}
\alias{aggregate_quantile}
\alias{aggregate_fv}
\title{Defunct Functions}
\usage{
aggregate_quantile(...)

aggregate_fv(...)
}
\description{
Functions mentioned in hard-copy journals, but later \link[base]{.Defunct}.
}
\keyword{internal}
