% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Defunct.R
\name{defunct}
\alias{defunct}
\alias{aggregate_quantile}
\alias{summary_fv}
\title{Defunct Functions}
\usage{
aggregate_quantile(
  ...,
  new = "<groupedHyperframe> |> quantile() |> aggregate()"
)

summary_fv(..., new = "<hyperframe> |> keyval() |> cumvtrapz()")
}
\arguments{
\item{...}{previous parameters}

\item{new}{\link[base]{character} scalar, see function \link[base]{.Defunct}.}
}
\description{
Functions mentioned in hard-copy journals, but later \link[base]{.Defunct}.
}
\keyword{internal}
