% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_marks.R
\name{density_marks}
\alias{density_marks}
\alias{density_marks.ppp}
\alias{density_marks.ppplist}
\title{Kernel \link[stats]{density} of \link[spatstat.geom]{marks}}
\usage{
density_marks(x, ...)

\method{density_marks}{ppp}(x, ...)

\method{density_marks}{ppplist}(x, ...)
}
\arguments{
\item{x}{a \link[spatstat.geom]{ppp.object}, or \code{'ppplist'}}

\item{...}{additional parameters of the function \link[stats]{density.default}}
}
\value{
The function \code{\link[=density_marks]{density_marks()}} returns
a \link[base]{numeric} \link[base]{vector}.
}
\description{
Kernel \link[stats]{density} of \link[spatstat.geom]{marks}.
}
\details{
The function \code{\link[=density_marks]{density_marks()}} finds
the kernel density of the \link[spatstat.geom]{marks}.
}
\note{
The function \code{spatstat.explore::density.ppp()} is for \eqn{x}- and \eqn{y}-\link[spatstat.geom]{coords} only!
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
