% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvlist.R
\name{fvlist}
\alias{fvlist}
\alias{key1val.fvlist}
\alias{cumtrapz.fvlist}
\title{Operations on a \link[base]{list} of \link[spatstat.explore]{fv.object}s}
\usage{
key1val.fvlist(X, check = TRUE, ...)

cumtrapz.fvlist(
  X,
  check = TRUE,
  mc.cores = switch(.Platform$OS.type, windows = 1L, detectCores()),
  ...
)
}
\arguments{
\item{X}{a \link[base]{list} of \link[spatstat.explore]{fv.object}s}

\item{check}{\link[base]{logical} scalar, an option to suppress
function \code{\link[=check_fvlist]{check_fvlist()}} in a batch process.
Default \code{TRUE}}

\item{...}{additional parameters, currently not in use}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}.
Default is 1L on Windows, or \link[parallel]{detectCores} on Mac.
CRAN requires \code{mc.cores <= 2L} in examples.}
}
\value{
All functions return a \link[base]{double} \link[base]{matrix}.
}
\description{
Operations on a \link[base]{list} of \link[spatstat.explore]{fv.object}s
}
\details{
Function \code{\link[=key1val.fvlist]{key1val.fvlist()}} gathers the primary outcome (via function \code{\link[=key1.fv]{key1.fv()}})
of the \link[spatstat.explore]{fv.object}s.

Function \code{\link[=cumtrapz.fvlist]{cumtrapz.fvlist()}} is a batch process of function \code{\link[=cumtrapz.fv]{cumtrapz.fv()}}.
}
\keyword{internal}
