% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_ppp.R
\name{get_nested}
\alias{get_nested}
\alias{get_nested_factors}
\title{Get Nested Levels}
\usage{
get_nested(group)

get_nested_factors(group, data)
}
\arguments{
\item{group}{a \link[base]{language} object, (nested) grouping structure}

\item{data}{\link[base]{data.frame} or \link[spatstat.geom]{hyperframe}}
}
\description{
Get Nested Levels
}
\keyword{internal}
