% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{groupedHyperframe-package}
\alias{groupedHyperframe}
\alias{groupedHyperframe-package}
\title{groupedHyperframe: Grouped Hyper Data Frame}
\description{
An S3 class 'groupedHyperframe' that inherits from the hyper data frame class from the 'spatstat' packages. Batch process and aggregation of the hyper column(s) over a nested grouping structure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tingtingzhan/groupedHyperframe}
  \item \url{https://tingtingzhan.quarto.pub/groupedhyperframe}
  \item \url{https://tingtingzhan-groupedhyperframe.netlify.app}
}

}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

}
