% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{groupedHyperframe-package}
\alias{groupedHyperframe-package}
\title{groupedHyperframe: Grouped Hyperframe: an Extension of hyperframe Object in Package spatstat.geom.}
\description{
An S3 class groupedHyperframe that inherits from hyperframe (defined by package spatstat.geom). Batch processes on ppp-hypercolumn. Aggregation of fv-hypercolumn(s) and numeric-hypercolumn(s) over nested grouping structure.
}
\references{
To be added.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

Authors:
\itemize{
  \item Inna Chervoneva \email{Inna.Chervoneva@jefferson.edu} (\href{https://orcid.org/0000-0002-9104-4505}{ORCID})
}

}
