% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{groupedHyperframe-package}
\alias{groupedHyperframe}
\alias{groupedHyperframe-package}
\title{groupedHyperframe: Grouped Hyper Data Frame}
\description{
An S3 class 'groupedHyperframe' that inherits from hyper data frame. Batch processes and aggregation of hyper column(s) over a nested grouping structure.
}
\note{
Help files of individual functions are intentionally suppressed in the \code{pdf} manual.
Users are encouraged to get started with the vignettes.

\url{https://tingtingzhan.quarto.pub/groupedhyperframe}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tingtingzhan/groupedHyperframe}
  \item \url{https://tingtingzhan.quarto.pub/groupedhyperframe}
}

}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

}
