% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{groupGeneric_psp}
\alias{groupGeneric_psp}
\alias{Math.psp}
\alias{Summary.psp}
\title{\link[base]{groupGeneric} of \link[spatstat.geom]{psp.object}}
\usage{
\method{Math}{psp}(x, ...)

\method{Summary}{psp}(..., na.rm = FALSE)
}
\arguments{
\item{x}{a \link[spatstat.geom]{psp.object}}

\item{...}{additional parameters for the \link[base]{groupGeneric} functions}

\item{na.rm}{\link[base]{logical} scalar}
}
\value{
Function \code{\link[=Math.psp]{Math.psp()}} returns a \link[spatstat.geom]{psp.object}.

Function \code{\link[=Summary.psp]{Summary.psp()}} ..
}
\description{
...
}
\details{
Function \code{\link[=Math.psp]{Math.psp()}} performs \code{Math} operations on the \link[base]{numeric} \link[spatstat.geom]{marks}
of a \link[spatstat.geom]{psp.object}.

Function \code{\link[=Summary.psp]{Summary.psp()}} ..
}
\keyword{internal}
