% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{groupGeneric_tess}
\alias{groupGeneric_tess}
\alias{Math.tess}
\alias{Summary.tess}
\title{\link[base]{groupGeneric} of \link[spatstat.geom]{tess}ellation}
\usage{
\method{Math}{tess}(x, ...)

\method{Summary}{tess}(..., na.rm = FALSE)
}
\arguments{
\item{x}{a \link[spatstat.geom]{tess}ellation}

\item{...}{additional parameters for the \link[base]{groupGeneric} functions}

\item{na.rm}{\link[base]{logical} scalar}
}
\value{
Function \code{\link[=Math.tess]{Math.tess()}} returns a \link[spatstat.geom]{ppp.object}.

Function \code{\link[=Summary.tess]{Summary.tess()}} ..
}
\description{
...
}
\details{
Function \code{\link[=Math.tess]{Math.tess()}} performs \code{Math} operations on the \link[base]{numeric} \link[spatstat.geom]{marks}
of a \link[spatstat.geom]{tess}ellation.

Function \code{\link[=Summary.tess]{Summary.tess()}} ..
}
\keyword{internal}
