% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvlist.R
\name{illegal2}
\alias{illegal2}
\alias{illegal2theo}
\title{Tentative Fix for Illegal Function Value in \code{fv.object}}
\usage{
illegal2theo(X)
}
\arguments{
\item{X}{an \link[spatstat.explore]{fv.object}}
}
\description{
Tentative Fix for Illegal Function Value in \code{fv.object}
}
\keyword{internal}
