% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.fv.R
\name{interpolation_fv}
\alias{interpolation_fv}
\alias{approxfun.fv}
\alias{splinefun.fv}
\title{\link[stats]{approxfun} of \link[spatstat.explore]{fv.object}}
\usage{
approxfun.fv(x, key = fvnames(fv, a = ".y"), ...)

splinefun.fv(x, key = fvnames(fv, a = ".y"), ...)
}
\arguments{
\item{x}{an \link[spatstat.explore]{fv.object}}

\item{key}{\link[base]{character} scalar, see function \code{\link[=keyval.fv]{keyval.fv()}}}

\item{...}{additional parameters of functions
\link[stats]{approxfun} and \link[stats]{splinefun}.}
}
\description{
\link[stats]{approxfun} of \link[spatstat.explore]{fv.object}
}
\keyword{internal}
