% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorlist.R
\name{is.vectorlist}
\alias{is.vectorlist}
\title{\link{is.vectorlist}}
\usage{
is.vectorlist(
  x,
  mode = c("any", "logical", "integer", "numeric", "double", "character")
)
}
\arguments{
\item{x}{an \link[spatstat.geom]{anylist}}

\item{mode}{\link[base]{character} scalar other than \code{'any'}, \code{'complex'} and '\code{raw}',
see function \link[base]{is.vector}}
}
\description{
\link{is.vectorlist}
}
\keyword{internal}
