% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorlist.R
\name{is.vectorlist}
\alias{is.vectorlist}
\title{\link{is.vectorlist}}
\usage{
is.vectorlist(
  x,
  mode = c("logical", "integer", "numeric", "double", "character")
)
}
\arguments{
\item{x}{a \link[stats]{listof}}

\item{mode}{\link[base]{character} scalar other than \code{'any'}, \code{'complex'} and '\code{raw}',
see function \link[base]{is.vector}}
}
\description{
\link{is.vectorlist}
}
\examples{
spatstat.data::Kovesi$values |>
 is.vectorlist(mode = 'character') |>
 stopifnot()
spatstat.data::Kovesi$values |>
 is.vectorlist(mode = 'numeric')
}
\keyword{internal}
