% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorlist.R
\name{is.vectorlist}
\alias{is.vectorlist}
\title{\link{is.vectorlist}}
\usage{
is.vectorlist(x, ...)
}
\arguments{
\item{x}{an \link[spatstat.geom]{anylist}}

\item{...}{additional parameter of \link[base]{is.vector}, currently only \code{mode}}
}
\description{
\link{is.vectorlist}
}
\keyword{internal}
