% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerndens.R
\name{kerndens}
\alias{kerndens}
\alias{kerndens.numeric}
\alias{kerndens.ppp}
\alias{kerndens.ppplist}
\alias{kerndens.anylist}
\alias{kerndens.hyperframe}
\title{Kernel Density via \link[stats]{density.default}}
\usage{
kerndens(x, ...)

\method{kerndens}{numeric}(x, ...)

\method{kerndens}{ppp}(x, ...)

\method{kerndens}{ppplist}(x, ...)

\method{kerndens}{anylist}(x, ...)

\method{kerndens}{hyperframe}(x, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{additional parameters of the function \link[stats]{density.default}}
}
\value{
The function \code{\link[=kerndens]{kerndens()}} returns
a \link[base]{numeric} \link[base]{vector}.
}
\description{
Kernel density estimates only.
}
\details{
The function \code{\link[=kerndens]{kerndens()}} finds
the estimated density values,
i.e., element \verb{$y} of the function \link[stats]{density.default} return.
}
\note{
Do \emph{not} overwrite function \code{spatstat.explore::density.ppp()} !!
}
\references{
\url{https://tingtingzhan.quarto.pub/groupedhyperframe/topics.html}
}
\keyword{internal}
