% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key1.R
\name{key1}
\alias{key1}
\alias{keyval.fv}
\alias{trapz.fv}
\alias{vtrapz.fv}
\alias{cumtrapz.fv}
\alias{cumvtrapz.fv}
\title{Black-Solid-Curve in \link[spatstat.explore]{plot.fv}}
\usage{
keyval.fv(x, key = fvnames(x, a = ".y"))

trapz.fv(x, key = fvnames(x, a = ".y"))

vtrapz.fv(x, key = fvnames(x, a = ".y"))

cumtrapz.fv(x, key = fvnames(x, a = ".y"))

cumvtrapz.fv(x, key = fvnames(x, a = ".y"))
}
\arguments{
\item{x}{an \link[spatstat.explore]{fv.object}}

\item{key}{\link[base]{character} scalar, default value is \code{spatstat.explore::fvnames(x, a = '.y')},
to speed up batch processes.}
}
\value{
Function \code{\link[=keyval.fv]{keyval.fv()}} returns a \link[base]{numeric} \link[base]{vector}.

Functions \code{\link[=trapz.fv]{trapz.fv()}} and \code{\link[=vtrapz.fv]{vtrapz.fv()}} return a \link[base]{numeric} scalar.

Functions \code{\link[=cumtrapz.fv]{cumtrapz.fv()}} and \code{\link[=cumvtrapz.fv]{cumvtrapz.fv()}} return a \link[base]{numeric} \link[base]{vector}.
}
\description{
Name and value of the \emph{black solid curve} as shown in \link[spatstat.explore]{plot.fv},
i.e., the primary outcome of an \link[spatstat.explore]{fv.object}.
}
\details{
Function \code{\link[=keyval.fv]{keyval.fv()}} finds the value of the (primary) outcome
of an \link[spatstat.explore]{fv.object}.

Functions \code{\link[=trapz.fv]{trapz.fv()}} and \code{\link[=cumtrapz.fv]{cumtrapz.fv()}}
obtain the (cumulative)
\link[pracma]{trapz}oidal integration of the area under the primary outcome
of a function value \link[spatstat.explore]{fv.object}.
}
\keyword{internal}
