% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{kmeans}
\alias{kmeans}
\alias{.kmeans}
\alias{.kmeans.ppp}
\title{Extension of \link[stats]{kmeans}}
\usage{
.kmeans(x, formula, ...)

\method{.kmeans}{ppp}(x, formula, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{formula}{\link[stats]{formula}}

\item{...}{additional parameters of function \link[stats]{kmeans}}
}
\value{
Function \code{\link[=.kmeans]{.kmeans()}} returns a \link[stats]{kmeans} object.
}
\description{
Extension of \link[stats]{kmeans}
}
\examples{
library(spatstat.data)
library(spatstat.geom)
shapley |> plot()
km = shapley |> .kmeans(formula = ~ x + y + Mag, centers = 3L)
km |> class()
}
\keyword{internal}
