% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fvlist.R
\name{lastLegal}
\alias{lastLegal}
\title{Last Legal Index}
\usage{
lastLegal(v)
}
\arguments{
\item{v}{\link[base]{double} \link[base]{vector}, return of function \code{\link[=key1val.fv]{key1val.fv()}}}
}
\description{
Last Legal Index
}
\details{
Legal, meaning not \code{0}, not \code{NaN} and not \code{Inf}.
}
\examples{
c(1, 1, 1) |> lastLegal()
c(1, 1, 1, NaN) |> lastLegal()
c(1, 1, 1, NaN, 1, 0, Inf) |> lastLegal()
# all return `3`
}
\keyword{internal}
