% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess.fv.R
\name{loess.fv}
\alias{loess.fv}
\title{\link[stats]{loess} of \link[spatstat.explore]{fv.object}}
\usage{
loess.fv(x, key = fvnames(fv, a = ".y"), ...)
}
\arguments{
\item{x}{an \link[spatstat.explore]{fv.object}}

\item{key}{\link[base]{character} scalar, see function \code{\link[=keyval.fv]{keyval.fv()}}}

\item{...}{additional parameters of functions
\link[stats]{loess}, except for \code{formula} and \code{data}}
}
\description{
\link[stats]{loess} of \link[spatstat.explore]{fv.object}
}
\keyword{internal}
