% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_S3.R
\name{log_ppp}
\alias{log_ppp}
\alias{log.ppp}
\alias{log1p.ppp}
\title{log.ppp}
\usage{
\method{log}{ppp}(x, base = exp(1))

\method{log1p}{ppp}(x)
}
\arguments{
\item{x}{a \link[spatstat.geom]{ppp.object}}

\item{base}{\link[base]{numeric} scalar}
}
\value{
Function \code{\link[=log.ppp]{log.ppp()}} returns a \link[spatstat.geom]{ppp.object}.
}
\description{
...
}
\details{
Function \code{\link[=log.ppp]{log.ppp()}} takes a \link[base]{log} of continuous marks
of a \link[spatstat.geom]{ppp.object}.
}
\examples{
data(longleaf, package = 'spatstat.data')
longleaf |> plot()
longleaf |> log() |> plot()
}
\keyword{internal}
