% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_identical_by.R
\name{mc_identical_by}
\alias{mc_identical_by}
\title{Check if Data Column is Identical by Grouping}
\usage{
mc_identical_by(data, f, mc.cores = getOption("mc.cores"), ...)
}
\arguments{
\item{data}{\link[base]{data.frame}}

\item{f}{\link[base]{factor}}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}.
Default is 1L on Windows, or \link[parallel]{detectCores} on Mac.
CRAN requires \code{mc.cores <= 2L} in examples.}

\item{...}{additional parameters, currently not in use}
}
\description{
Check if Data Column is Identical by Grouping
}
\note{
Function \link[stats]{aggregate.data.frame} does not do parallel computing.

Function \code{collapse::collap} does not support \link[survival]{Surv} column.
}
\keyword{internal}
