% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette.R
\name{methods2kable}
\alias{methods2kable}
\title{\code{S3} methods table in vignette}
\usage{
methods2kable(class, package, ...)
}
\arguments{
\item{class}{\link[base]{character} scalar}

\item{package}{\link[base]{character} scalar}

\item{...}{additional parameters of function \link[utils]{methods}}
}
\value{
Function \code{\link[=methods2kable]{methods2kable()}} returns a \link[base]{data.frame}.
}
\description{
\code{S3} methods table in vignette
}
\keyword{internal}
