% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_S3.R
\name{na_fail_ppp}
\alias{na_fail_ppp}
\alias{na.omit.ppp}
\alias{na.exclude.ppp}
\title{Handle Missing \link[spatstat.geom]{marks} in \link[spatstat.geom]{ppp.object}.}
\usage{
\method{na.omit}{ppp}(object, ...)

\method{na.exclude}{ppp}(object, ...)
}
\arguments{
\item{object}{a \link[spatstat.geom]{ppp.object}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=na.omit.ppp]{na.omit.ppp()}} returns a \link[spatstat.geom]{ppp.object}.
}
\description{
Handle Missing \link[spatstat.geom]{marks} in \link[spatstat.geom]{ppp.object}.
}
\details{
Function \code{\link[=na.omit.ppp]{na.omit.ppp()}} omits missing \link[spatstat.geom]{marks} in a \link[spatstat.geom]{ppp.object}.
}
\note{
tzh suppose missing \verb{$x} and \verb{$y} are
forbidden in \link[spatstat.geom]{ppp.object} anyway.
}
\keyword{internal}
