% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_ppplist.R
\name{op_ppplist}
\alias{op_ppplist}
\title{Batch Operations of \code{'ppplist'} Object}
\usage{
op_ppplist(x, op, mc.cores = getOption("mc.cores"), ...)
}
\arguments{
\item{x}{a \code{'ppplist'} object}

\item{op}{workhorse \link[base]{function}, either \code{\link[=ppp2fv]{ppp2fv()}} or \code{\link[=ppp2dist]{ppp2dist()}}}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}.
Default is 1L on Windows, or \link[parallel]{detectCores} on Mac.}

\item{...}{additional parameters of workhorse functions
\code{\link[=ppp2fv]{ppp2fv()}} or \code{\link[=ppp2dist]{ppp2dist()}}}
}
\value{
Function \code{\link[=op_ppplist]{op_ppplist()}} returns a \link[stats]{listof}
\itemize{
\item function \code{\link[=ppp2fv]{ppp2fv()}} returns, if \code{op = ppp2fv}.
\item function \code{\link[=ppp2dist]{ppp2dist()}} returns, if \code{op = ppp2dist}.
}
}
\description{
Batch operations of function \code{\link[=ppp2fv]{ppp2fv()}} or \code{\link[=ppp2dist]{ppp2dist()}}, for a \code{'ppplist'} input.
}
\details{
Function \code{\link[=op_ppplist]{op_ppplist()}} is a \pkg{parallel} batch process of
the workhorse function \code{\link[=ppp2fv]{ppp2fv()}} or \code{\link[=ppp2dist]{ppp2dist()}}.
}
\examples{
\dontshow{options(mc.cores = 1L)}
library(spatstat.data)
library(spatstat.geom) # for ?spatstat.geom::split.ppp
library(spatstat.explore) # for ?spatstat.explore::Emark, etc.

Vc = with(shapley$marks, expr = {
 cut.default(V, breaks = quantile(V, probs = c(0, 1/3, 2/3, 1)), labels = c('L', 'M', 'H'))
})
\donttest{
shapley |> 
 subset.ppp(select = c('Mag', 'SigV')) |>
 split.ppp(f = Vc) |>
 op_ppplist(op = ppp2fv, fn = markcorr)
}

}
\keyword{internal}
