% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_spatial.R
\name{pairwise_cor_spatial}
\alias{pairwise_cor_spatial}
\alias{pairwise_cor_spatial.ppp}
\title{Pairwise Tjostheim's Coefficient}
\usage{
pairwise_cor_spatial(x, ...)

\method{pairwise_cor_spatial}{ppp}(x, formula, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{additional parameters, currently not in use}

\item{formula}{\link[stats]{formula}}
}
\value{
Function \code{\link[=pairwise_cor_spatial]{pairwise_cor_spatial()}} returns a \link[stats]{dist} object.
}
\description{
Pairwise Tjostheim's Coefficient
}
\examples{
data(finpines, package = 'spatstat.data')
(r = finpines |> pairwise_cor_spatial.ppp(formula = ~ diameter + height))
r |> as.matrix()
}
\keyword{internal}
