% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmean.R
\name{pmean}
\alias{pmean}
\alias{pmedian}
\title{Parallel \link[base]{mean} and \link[stats]{median}}
\usage{
pmean(..., na.rm = TRUE)

pmedian(..., na.rm = TRUE)
}
\arguments{
\item{...}{\link[base]{numeric} \link[base]{vector}s}

\item{na.rm}{\link[base]{logical} scalar, default \code{TRUE}}
}
\value{
Functions \code{\link[=pmean]{pmean()}} and \code{\link[=pmedian]{pmedian()}} return a \link[base]{numeric} \link[base]{vector}.
}
\description{
Parallel \link[base]{mean} and \link[stats]{median}
}
\note{
Functions \code{\link[=pmean]{pmean()}} and \code{\link[=pmedian]{pmedian()}} mimic functions \link[base]{pmax.int} and \link[base]{pmin.int}.
They are written in a very naive way.
The pipeline \code{cbind() |> rowMeans()} is extremely slow.
}
\examples{
pmean(1:3, 11:13, 21:23)
}
\keyword{internal}
