% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_ppp.R
\name{ppp2.}
\alias{ppp2.}
\alias{ppp2dist}
\alias{ppp2fv}
\title{Operations on \link[spatstat.geom]{ppp.object}}
\usage{
ppp2dist(x, fun, ...)

ppp2fv(x, fun, ...)
}
\arguments{
\item{x}{a \link[spatstat.geom]{ppp.object}}

\item{fun}{a distance \link[base]{function},
or a \link[base]{function} that returns an \link[spatstat.explore]{fv.object},
see \strong{Details}}

\item{...}{additional parameters of function \code{fun}}
}
\value{
Function \code{\link[=ppp2fv]{ppp2fv()}} returns a \link[stats]{listof}
\link[spatstat.explore]{fv.object}s.

Function \code{\link[=ppp2dist]{ppp2dist()}} returns a \link[stats]{listof}
\link[base]{double} \link[base]{vector}s.
}
\description{
Create
\link[spatstat.explore]{fv.object}s
or
distances
from a \link[spatstat.geom]{ppp.object}.
}
\keyword{internal}
