% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmax.R
\name{rmax}
\alias{rmax}
\alias{rmax_}
\alias{rmax_.ppp}
\alias{rmax_.ppplist}
\alias{rmax_.hyperframe}
\title{Default \eqn{r_\text{max}} of Function \link[spatstat.explore]{markcorr}}
\usage{
rmax_(X, fun, ...)

\method{rmax_}{ppp}(X, fun = "K", ...)

\method{rmax_}{ppplist}(X, fun = "K", ...)

\method{rmax_}{hyperframe}(X, fun = "K", ...)
}
\arguments{
\item{X}{a \link[spatstat.geom]{ppp.object} with one \link[base]{numeric} \link[spatstat.geom]{marks}}

\item{...}{additional parameters, currently of no use}
}
\description{
Default \eqn{r_\text{max}} of Function \link[spatstat.explore]{markcorr}
}
\details{
Code snippet stolen from
function \link[spatstat.explore]{markcorr} (\code{fun = 'K'})
and making use of undocumented functions
\link[spatstat.explore]{rmax.rule} and \link[spatstat.geom]{handle.r.b.args}
}
\examples{
library(spatstat.data)
# ppp
spruces |> rmax_(fun = 'K') # rectangle window
urkiola |> rmax_(fun = 'K') # polygonal boundary 
swedishpines |> rmax_(fun = 'K') # not-marked, exception handling
# hyperframe
flu |> rmax_(fun = 'K') # rectangle window
cetaceans |> rmax_(fun = 'K') # polygonal boundary, marked and unmarked
pyramidal |> rmax_(fun = 'K') # not-marked
}
\keyword{internal}
