% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmax.R
\name{rmax}
\alias{rmax}
\alias{rmax_}
\alias{rmax_.ppp}
\alias{rmax_.ppplist}
\alias{rmax_.hyperframe}
\title{Default \eqn{r_\text{max}} of Various Functions in Package \CRANpkg{spatstat.explore}}
\usage{
rmax_(X, ...)

\method{rmax_}{ppp}(X, fun, i, j, ...)

\method{rmax_}{ppplist}(X, ...)

\method{rmax_}{hyperframe}(X, ...)
}
\arguments{
\item{X}{a \link[spatstat.geom]{ppp.object} with one \link[base]{numeric} or multi-type \link[spatstat.geom]{marks}}

\item{...}{additional parameters, currently of no use}

\item{fun}{\link[base]{character} scalar, see (the un-documented) function \link[spatstat.explore]{rmax.rule}}

\item{i, j}{\link[base]{character} scalars, see functions
\link[spatstat.explore]{Gcross},
\link[spatstat.explore]{Kcross},
\link[spatstat.explore]{Jcross}, etc.}
}
\description{
Default \eqn{r_\text{max}} of Various Functions in Package \CRANpkg{spatstat.explore}
}
\examples{
# ppp, fun = 'K'
spatstat.data::spruces |> rmax_(fun = 'K') # rectangle window
spatstat.data::urkiola |> rmax_(fun = 'K') # polygonal boundary 
spatstat.data::swedishpines |> rmax_(fun = 'K') # not-marked, exception handling

# hyperframe, fun = 'K'
spatstat.data::flu |> rmax_(fun = 'K') # rectangle window
spatstat.data::cetaceans |> rmax_(fun = 'K') # polygonal boundary, marked and unmarked
spatstat.data::pyramidal |> rmax_(fun = 'K') # not-marked

# hyperframe, fun = 'G'
spatstat.data::flu |> rmax_(fun = 'G')
flu0 = spatstat.data::flu |> subset(subset = (stain == 'M2-M1'))
flu0 |> rmax_(fun = 'G', i = 'M1', j = 'M2')
flu0 |> rmax_(fun = 'G', i = 'M2', j = 'M1')
}
\keyword{internal}
