% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmax.R
\name{rmax}
\alias{rmax}
\alias{rmax_}
\alias{rmax_.ppp}
\alias{rmax_.ppplist}
\alias{rmax_.hyperframe}
\title{Default \eqn{r_\text{max}} of Various Functions in Package \CRANpkg{spatstat.explore}}
\usage{
rmax_(X, ...)

\method{rmax_}{ppp}(X, fun, i, j, ...)

\method{rmax_}{ppplist}(X, ...)

\method{rmax_}{hyperframe}(X, ...)
}
\arguments{
\item{X}{a \link[spatstat.geom]{ppp.object} with one \link[base]{numeric} or multi-type \link[spatstat.geom]{marks}}

\item{...}{additional parameters, currently of no use}

\item{fun}{\link[base]{character} scalar, see (the un-documented) function \link[spatstat.explore]{rmax.rule}}

\item{i, j}{\link[base]{character} scalars, see functions
\link[spatstat.explore]{Gcross},
\link[spatstat.explore]{Kcross},
\link[spatstat.explore]{Jcross}, etc.}
}
\description{
Default \eqn{r_\text{max}} of Various Functions in Package \CRANpkg{spatstat.explore}
}
\keyword{internal}
