% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_kmeans.R
\name{split.pppkm}
\alias{split.pppkm}
\title{\code{split.pppkm}}
\usage{
\method{split}{pppkm}(x, f = attr(x, which = "f", exact = TRUE), ...)
}
\arguments{
\item{x}{a \code{'pppkm'} object, returned from function \code{\link[=.kmeans.ppp]{.kmeans.ppp()}}}

\item{f}{\link[base]{factor}, default is \code{attr(x,'f')}}

\item{...}{additional parameters, currently no use}
}
\value{
Function \code{\link[=split.pppkm]{split.pppkm()}} returns a \code{'splitppp'} object from the
workhorse function \link[spatstat.geom]{split.ppp}.
}
\description{
\code{split.pppkm}
}
