% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_kmeans.R
\name{split_kmeans}
\alias{split_kmeans}
\alias{split_kmeans.default}
\alias{split_kmeans.listof}
\alias{split_kmeans.hyperframe}
\title{\link[base]{split} by \link[stats]{kmeans} Clustering}
\usage{
split_kmeans(x, ...)

\method{split_kmeans}{default}(x, ...)

\method{split_kmeans}{listof}(x, ...)

\method{split_kmeans}{hyperframe}(x, ...)
}
\arguments{
\item{x}{see \strong{Usage}}

\item{...}{additional parameters of function \link[stats]{kmeans}}
}
\description{
\link[base]{split} by \link[stats]{kmeans} clustering
}
\note{
Function \code{\link[=split_kmeans.default]{split_kmeans.default()}} is supposed to work with
\link[spatstat.geom]{ppp.object}.
}
\keyword{internal}
