% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvlist.R
\name{summary.fvlist}
\alias{summary.fvlist}
\title{summary.fvlist}
\usage{
\method{summary}{fvlist}(
  object,
  data.name = deparse1(substitute(object)),
  rmax,
  mc.cores = getOption("mc.cores"),
  ...
)
}
\arguments{
\item{object}{an \code{'fvlist'}}

\item{data.name}{\link[base]{character} scalar}

\item{rmax}{(optional) \link[base]{numeric} scalar, user \eqn{r_\text{max}}}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}.
Default is 1L on Windows, or \link[parallel]{detectCores} on Mac.}

\item{...}{additional parameters, currently of no use}
}
\description{
summary.fvlist
}
\keyword{internal}
