% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surv.R
\name{summary.Surv}
\alias{summary.Surv}
\title{Summary Information of \link[survival]{Surv} Object}
\usage{
\method{summary}{Surv}(object, ...)
}
\arguments{
\item{object}{a \link[survival]{Surv} object}

\item{...}{additional parameters, currently of no use}
}
\description{
Summary Information of \link[survival]{Surv} Object
}
\details{
Otherwise dispatch to \link[base]{summary.default}.
}
\examples{
aml2 = survival::aml |>
 within.data.frame(expr = {
  os = survival::Surv(time = time, event = status)
  time = status = NULL
 })
summary(aml2$os)
summary(aml2)

heart2 = survival::heart |>
 within.data.frame(expr = {
  os = survival::Surv(time = start, time2 = stop, event = event)
  start = stop = event = NULL
 })
summary(heart2$os)
summary(heart2)

}
\keyword{internal}
