% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superimpose.hyperframe.R
\name{superimpose.hyperframe}
\alias{superimpose.hyperframe}
\title{superimpose.hyperframe}
\usage{
\method{superimpose}{hyperframe}(...)
}
\arguments{
\item{...}{one or more \link[spatstat.geom]{hyperframe}s}
}
\value{
Function \code{\link[=superimpose.hyperframe]{superimpose.hyperframe()}} returns a \link[spatstat.geom]{hyperframe}.
}
\description{
superimpose.hyperframe
}
\keyword{internal}
