% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtrapz.R
\name{visualize_vtrapz.fv}
\alias{visualize_vtrapz.fv}
\title{Visualize \code{\link[=vtrapz]{vtrapz()}} and \code{\link[=cumvtrapz]{cumvtrapz()}} of \link[spatstat.explore]{fv.object}}
\usage{
visualize_vtrapz.fv(x)
}
\arguments{
\item{x}{an \link[spatstat.explore]{fv.object}}
}
\description{
Visualize \code{\link[=vtrapz]{vtrapz()}} and \code{\link[=cumvtrapz]{cumvtrapz()}} of \link[spatstat.explore]{fv.object}
}
\examples{
spatstat.data::spruces |>
 spatstat.explore::Emark() |>
 visualize_vtrapz.fv() + ggplot2::theme_minimal()

spatstat.data::spruces |>
 spatstat.explore::Vmark() |>
 visualize_vtrapz.fv() + ggplot2::theme_minimal()

}
\keyword{internal}
