% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtrapz.R
\name{visualize_vtrapz}
\alias{visualize_vtrapz}
\title{Visualize \code{\link[=vtrapz]{vtrapz()}} and \code{\link[=cumvtrapz]{cumvtrapz()}}}
\usage{
visualize_vtrapz(x, y)
}
\arguments{
\item{x, y}{\link[base]{numeric} \link[base]{vector}s}
}
\description{
Visualize \code{\link[=vtrapz]{vtrapz()}} and \code{\link[=cumvtrapz]{cumvtrapz()}}
}
\keyword{internal}
