% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtrapz.R
\name{visualize_vtrapz}
\alias{visualize_vtrapz}
\title{Visualize \code{\link[=vtrapz]{vtrapz()}} and \code{\link[=cumvtrapz]{cumvtrapz()}}}
\usage{
visualize_vtrapz(x, y, draw.v = TRUE, draw.cumv = TRUE, draw.rect = TRUE)
}
\arguments{
\item{x, y}{\link[base]{numeric} \link[base]{vector}s}

\item{draw.v}{\link[base]{logical} scalar, whether to plot \code{\link[=vtrapz]{vtrapz()}}, default \code{TRUE}}

\item{draw.cumv}{\link[base]{logical} scalar, whether to plot \code{\link[=cumvtrapz]{cumvtrapz()}}, default \code{TRUE}}

\item{draw.rect}{\link[base]{logical} scalar, whether to plot the rectangle, default \code{TRUE}}
}
\description{
Visualize \code{\link[=vtrapz]{vtrapz()}} and \code{\link[=cumvtrapz]{cumvtrapz()}}
}
\keyword{internal}
