% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtrapz.R
\name{vtrapz}
\alias{vtrapz}
\alias{cumvtrapz}
\title{(Cumulative) Average Vertical Height of Trapezoidal Integration}
\usage{
vtrapz(x, ...)

cumvtrapz(x, ...)
}
\arguments{
\item{x}{\link[base]{numeric} \link[base]{vector}}

\item{...}{additional parameters of function \link[pracma]{trapz} and \link[pracma]{cumtrapz}}
}
\description{
(Cumulative) trapezoidal integration divided by \eqn{x}-domain.
}
\note{
This is a tentative thought: the prefix \code{v} stands for 'vertical'.
}
\examples{
(x = seq.int(from = 10L, to = 20L, by = 2L))
set.seed(12); (y = rnorm(n = length(x), mean = 1, sd = .2))
plot(x, y, type = 'b')

pracma::trapz(x, y)
vtrapz(x, y)

pracma::cumtrapz(x, y)
cumvtrapz(x, y)

visualize_vtrapz(x, y) + ggplot2::theme_minimal()

}
\keyword{internal}
