% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtrapz.R
\name{vtrapz}
\alias{vtrapz}
\alias{cumvtrapz}
\title{(Cumulative) Trapezoidal Integration Divided by \eqn{x}-Domain}
\usage{
vtrapz(x, ...)

cumvtrapz(x, ...)
}
\arguments{
\item{x}{\link[base]{numeric} \link[base]{vector}}

\item{...}{additional parameters of function \link[pracma]{trapz} and \link[pracma]{cumtrapz}}
}
\description{
(Cumulative) trapezoidal integration divided by \eqn{x}-domain.
}
\note{
This is a tentative thought: the prefix \code{v} stands for 'vertical'.
}
\examples{
(x = seq.int(from = 10L, to = 20L, by = 2L))
set.seed(12); (y = rnorm(n = length(x), mean = 1, sd = .2))
plot(x, y, type = 'b')

pracma::trapz(x, y)
vtrapz(x, y)

pracma::cumtrapz(x, y)
cumvtrapz(x, y)

\donttest{
library(ggplot2)
library(geomtextpath)

ggplot() + 
 geom_path(mapping = aes(x = x, y = y), alpha = .5) +
 geom_rect(mapping = aes(xmin = min(x), xmax = max(x), ymin = 0, ymax = vtrapz(x, y)), alpha = .1) +
 geom_textpath(
  mapping = aes(x = x, y = vtrapz(x, y), label = 'Average Height of Trapezoidal Integration'),
  hjust = .1, text_only = TRUE, colour = 'red', fontface = 'bold'
 ) +
 geom_textpath(
  mapping = aes(
   x = x[-1L], y = cumvtrapz(x, y)[-1L], 
   label = 'Cumulative Average Height of Trapezoidal Integration'
  ),
  text_only = TRUE, colour = 'blue', fontface = 'bold'
 ) +
 scale_x_continuous(breaks = x, limits = c(9, 21)) + 
 ylim(c(0, 1.5)) + 
 theme_bw()
}
}
\keyword{internal}
